TelnyxAPI.Api.AccessControlIPs (telnyx_api v2.0.0)

API calls for all endpoints tagged AccessControlIPs.

Functions

create_access_control_ip(connection, access_control_ip_create, opts \\ [])

@spec create_access_control_ip(
  Tesla.Env.client(),
  TelnyxAPI.Model.AccessControlIpCreate.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.AccessControlIpResponse.t()}
  | {:error, Tesla.Env.t()}

Adds an access control IP This endpoint will be deprecated on 15th March 2022 and has been replaced by the following APIs, api.telnyx.com/v2/access_ip_address and api.telnyx.com/v2/access_ip_ranges. We recommend that Customers already using this endpoint transition to using: api.telnyx.com/v2/access_ip_address.)

Parameters

connection (TelnyxAPI.Connection): Connection to server

access_control_ip_create (AccessControlIpCreate): Parameters that can be defined during access control IP creation

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_access_control_ip(connection, id, opts \\ [])

@spec delete_access_control_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.AccessControlIpResponse.t()}
  | {:error, Tesla.Env.t()}

Deletes an access control IP This endpoint will be deprecated on 15th March 2022 and has been replaced by the following APIs, api.telnyx.com/v2/access_ip_address and api.telnyx.com/v2/access_ip_ranges. We recommend that Customers already using this endpoint transition to using: api.telnyx.com/v2/access_ip_address.)

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): access control IP ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

find_access_control_ips(connection, opts \\ [])

@spec find_access_control_ips(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.GetAllAccessControlIpResponse.t()}
  | {:error, Tesla.Env.t()}

List all access control IPs This endpoint will be deprecated on 15th March 2022 and has been replaced by the following APIs, api.telnyx.com/v2/access_ip_address and api.telnyx.com/v2/access_ip_ranges. We recommend that Customers already using this endpoint transition to using: api.telnyx.com/v2/access_ip_address.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketip_address_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter access control IPs via the ip_address sent.

Returns

on success {:error, Tesla.Env.t} on failure

get_access_control_ip(connection, id, opts \\ [])

@spec get_access_control_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.AccessControlIpResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve an access control IP This endpoint will be deprecated on 15th March 2022 and has been replaced by the following APIs, api.telnyx.com/v2/access_ip_address and api.telnyx.com/v2/access_ip_ranges. We recommend that Customers already using this endpoint transition to using: api.telnyx.com/v2/access_ip_address.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): access control IP ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_access_control_ip(connection, id, access_control_ip_update, opts \\ [])

@spec update_access_control_ip(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.AccessControlIpUpdate.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.AccessControlIpResponse.t()}
  | {:error, Tesla.Env.t()}

Update an access control IP This endpoint will be deprecated on 15th March 2022 and has been replaced by the following APIs, api.telnyx.com/v2/access_ip_address and api.telnyx.com/v2/access_ip_ranges. We recommend that Customers already using this endpoint transition to using: api.telnyx.com/v2/access_ip_address.)

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): access control IP ID

access_control_ip_update (AccessControlIpUpdate): Parameters to update on the access control IP record

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure