TelnyxAPI.Api.BulkCredentials (telnyx_api v2.0.0)

API calls for all endpoints tagged BulkCredentials.

Functions

bulk_credential_action(connection, action, filter_left_square_brackettag_right_square_bracket, opts \\ [])

@spec bulk_credential_action(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.BulkCredentialResponse.t()}
  | {:ok, nil}
  | {:error, Tesla.Env.t()}

Perform activate or deactivate action on all credentials filtered by the provided tag. Perform activate or deactivate action on all credentials filtered by the provided tag. Activate action will change the status to active, making it possible to connect calls with the credential. Deactivate action will change the status to inactive, making it impossible to connect calls with the credential.

Parameters

connection (TelnyxAPI.Connection): Connection to server

action (String.t): Identifies the action to be taken. Activate will change the status to active. Deactivate will change the status to inactive.

filter_left_square_brackettag_right_square_bracket (String.t): Filter by tag, required by bulk operations.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_bulk_telephony_credential(connection, bulk_credential_request, opts \\ [])

@spec create_bulk_telephony_credential(
  Tesla.Env.client(),
  TelnyxAPI.Model.BulkCredentialRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.BulkCredentialResponse.t()}
  | {:ok, nil}
  | {:error, Tesla.Env.t()}

Creates several credentials Creates several credentials in bulk.

Parameters

connection (TelnyxAPI.Connection): Connection to server

bulk_credential_request (BulkCredentialRequest): Requested parameters to create credentials on bulk

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_bulk_telephony_credential(connection, filter_left_square_brackettag_right_square_bracket, opts \\ [])

@spec delete_bulk_telephony_credential(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.BulkCredentialResponse.t()}
  | {:ok, nil}
  | {:error, Tesla.Env.t()}

Delete several credentials Delete several credentials in bulk.

Parameters

connection (TelnyxAPI.Connection): Connection to server

filter_left_square_brackettag_right_square_bracket (String.t): Filter by tag, required by bulk operations.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_bulk_telephony_credential(connection, filter_left_square_brackettag_right_square_bracket, bulk_credential_request, opts \\ [])

@spec update_bulk_telephony_credential(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.BulkCredentialRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.BulkCredentialResponse.t()}
  | {:ok, nil}
  | {:error, Tesla.Env.t()}

Update several credentials Update several credentials in bulk.

Parameters

connection (TelnyxAPI.Connection): Connection to server

filter_left_square_brackettag_right_square_bracket (String.t): Filter by tag, required by bulk operations.

bulk_credential_request (BulkCredentialRequest): Parameters to update credentials on bulk

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure