TelnyxAPI.Api.BusinessIdentity (telnyx_api v2.0.0)

API calls for all endpoints tagged BusinessIdentity.

Functions

create_business_identity(connection, create_or_update_business_identity_request, opts \\ [])

@spec create_business_identity(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateOrUpdateBusinessIdentityRequest.t(),
  keyword()
) :: {:ok, nil} | {:ok, TelnyxAPI.Model.Errors.t()} | {:error, Tesla.Env.t()}

Creates a Business Identity Creates a new Business Identity.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_or_update_business_identity_request (CreateOrUpdateBusinessIdentityRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_business_identity(connection, id, opts \\ [])

@spec delete_business_identity(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ShowBusinessIdentityResponse.t()}
  | {:error, Tesla.Env.t()}

Delete a Business Identity Deletes the Business Identity. This action will fail if any verification requests have been made for any associated Display Profile. Please contact support@telnyx.com in case you want to delete a Business Identity in that situation.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the Business Identity

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

display_business_identity(connection, id, opts \\ [])

@spec display_business_identity(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ShowBusinessIdentityResponse.t()}
  | {:error, Tesla.Env.t()}

Display the Business Identity

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the Business Identity

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_business_identities(connection, opts \\ [])

@spec list_business_identities(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListBusinessIdentitiesResponse.t()}
  | {:error, Tesla.Env.t()}

Lists the Business Identities owned by the current user/organization

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

update_business_identity(connection, id, create_or_update_business_identity_request, opts \\ [])

@spec update_business_identity(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.CreateOrUpdateBusinessIdentityRequest.t(),
  keyword()
) :: {:ok, nil} | {:ok, TelnyxAPI.Model.Errors.t()} | {:error, Tesla.Env.t()}

Update a Business Identity Update an existing Business Identity.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the Business Identity

create_or_update_business_identity_request (CreateOrUpdateBusinessIdentityRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure