TelnyxAPI.Api.CallControlApplications (telnyx_api v0.1.0)

API calls for all endpoints tagged CallControlApplications.

Functions

create_call_control_application(connection, create_call_control_application_request, opts \\ [])

Specs

create_call_control_application(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateCallControlApplicationRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.CallControlApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Create a call control application Create a call control application.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_call_control_application_request (CreateCallControlApplicationRequest): Create call control application request.

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure

delete_call_control_application(connection, id, opts \\ [])

Specs

delete_call_control_application(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.CallControlApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Delete a call control application Deletes a call control application.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure

list_call_control_applications(connection, opts \\ [])

Specs

list_call_control_applications(Tesla.Env.client(), keyword()) ::
  {:ok, TelnyxAPI.Model.ListCallControlApplicationsResponse.t()}
  | {:ok, nil}
  | {:error, Tesla.Env.t()}

List call control applications Return a list of call control applications.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketapplication_name_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): If present, applications with <code>application_name</code> containing the given value will be returned. Matching is not case-sensitive. Requires at least three characters.

:filter_left_square_bracketoutbound_periodoutbound_voice_profile_id_right_square_bracket (String.t): Identifies the associated outbound voice profile.

:sort (String.t): Specifies the sort order for results. By default sorting direction is ascending. To have the results sorted in descending order add the <code> -</code> prefix.<br/><br/> That is: <ul> <li> <code>connection_name</code>: sorts the result by the <code>connection_name</code> field in ascending order. </li> <li> <code>-connection_name</code>: sorts the result by the <code>connection_name</code> field in descending order. </li> </ul> <br/> If not given, results are sorted by <code>created_at</code> in descending order.
Returns

on success {:error, Tesla.Env.t} on failure

retrieve_call_control_application(connection, id, opts \\ [])

Specs

retrieve_call_control_application(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.CallControlApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a call control application Retrieves the details of an existing call control application.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure

update_call_control_application(connection, id, update_call_control_application_request, opts \\ [])

Specs

update_call_control_application(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateCallControlApplicationRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.CallControlApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Update a call control application Updates settings of an existing call control application.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

update_call_control_application_request (UpdateCallControlApplicationRequest): Update call control application request.

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure