TelnyxAPI.Api.ChannelZones (telnyx_api v2.0.0)

API calls for all endpoints tagged ChannelZones.

Functions

assign_phone_number(connection, channel_zone_id, body, opts \\ [])

@spec assign_phone_number(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ListChannelZonesByPhoneNumber.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.GcbPhoneNumber.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Assign a phone number to a channel zone You should own the phone number being assigned to the channel zone. Remember that you should reserve channels in this channel zone, otherwise you won't be able to receive incoming calls.

Parameters

connection (TelnyxAPI.Connection): Connection to server

channel_zone_id (String.t): Channel zone identifier

body (ListChannelZonesByPhoneNumber):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_channel_zone(connection, channel_zone_id, opts \\ [])

@spec get_channel_zone(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GcbChannelZone.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Get a specific channel zone Get a specific channel zone

Parameters

connection (TelnyxAPI.Connection): Connection to server

channel_zone_id (String.t): Channel zone identifier

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_channel_zones(connection, opts \\ [])

@spec get_channel_zones(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.GetGcbChannelZonesRequestResponse.t()}
  | {:error, Tesla.Env.t()}

Get list of channel zones List of channel zones with their countries

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

get_phone_numbers(connection, channel_zone_id, opts \\ [])

@spec get_phone_numbers(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetGcbPhoneNumbersRequestResponse.t()}
  | {:error, Tesla.Env.t()}

Get the list of phone numbers assigned to a channel zone Retrieve the assigned phone numbers in a channel zone. Phone numbers assigned to a channel zone can receive concurrent calls up to the quantity reserved in that channel zone. Additional concurrent calls are rejected with a busy signal.

Parameters

connection (TelnyxAPI.Connection): Connection to server

channel_zone_id (String.t): Channel zone identifier

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

patch_group(connection, channel_zone_id, body, opts \\ [])

@spec patch_group(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateChannelZone.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.GcbChannelZone.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Patch a channel zone Change the amount of reserved channels at a given channel zone

Parameters

connection (TelnyxAPI.Connection): Connection to server

channel_zone_id (String.t): Channel zone identifier

body (UpdateChannelZone):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

unassign_phone_number(connection, channel_zone_id, phone_number, opts \\ [])

@spec unassign_phone_number(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.Errors.t()} | {:error, Tesla.Env.t()}

Unassign a phone number from a channel zone

Parameters

connection (TelnyxAPI.Connection): Connection to server

channel_zone_id (String.t): Channel zone identifier

phone_number (String.t): The phone number to be looked up

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure