TelnyxAPI.Api.Credentials (telnyx_api v2.0.0)

API calls for all endpoints tagged Credentials.

Functions

create_telephony_credential(connection, telephony_credential_create_request, opts \\ [])

@spec create_telephony_credential(
  Tesla.Env.client(),
  TelnyxAPI.Model.TelephonyCredentialCreateRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.TelephonyCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

Create a credential Create a credential.

Parameters

connection (TelnyxAPI.Connection): Connection to server

telephony_credential_create_request (TelephonyCredentialCreateRequest): Parameters that can be defined during credential creation

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_telephony_credential(connection, id, opts \\ [])

@spec delete_telephony_credential(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.TelephonyCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

Delete a credential Delete an existing credential.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

find_telephony_credentials(connection, opts \\ [])

@spec find_telephony_credentials(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.GetAllTelephonyCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

List all credentials List all On-demand Credentials.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_brackettag_right_square_bracket (String.t): Filter by tag

:filter_left_square_bracketname_right_square_bracket (String.t): Filter by name

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter by status

:filter_left_square_bracketresource_id_right_square_bracket (String.t): Filter by resource_id

:filter_left_square_bracketsip_username_right_square_bracket (String.t): Filter by sip_username

Returns

on success {:error, Tesla.Env.t} on failure

get_telephony_credential(connection, id, opts \\ [])

@spec get_telephony_credential(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.TelephonyCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

Get a credential Get the details of an existing On-demand Credential.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_tags(connection, opts \\ [])

@spec list_tags(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListTagsResponse.t()}
  | {:error, Tesla.Env.t()}

List all tags Returns a list of tags used on Credentials

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

telephony_credential_action(connection, id, action, opts \\ [])

@spec telephony_credential_action(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.TelephonyCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

Perform activate or deactivate action on provided Credential Perform activate or deactivate action on provided Credential. Activate action will change the status to active, making it possible to connect calls with the credential. Deactivate action will change the status to inactive, making it impossible to connect calls with the credential.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

action (String.t): Identifies the action to be taken.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_telephony_credential(connection, id, telephony_credential_update_request, opts \\ [])

@spec update_telephony_credential(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.TelephonyCredentialUpdateRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.TelephonyCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

Update a credential Update an existing credential.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

telephony_credential_update_request (TelephonyCredentialUpdateRequest): Parameters that can be updated in a credential

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure