TelnyxAPI.Api.CustomerServiceRequest (telnyx_api v2.0.0)

API calls for all endpoints tagged CustomerServiceRequest.

Functions

get_customer_service_request(connection, customer_service_request_id, opts \\ [])

@spec get_customer_service_request(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ShowCustomerServiceRequest.t()}
  | {:ok, Map.t()}
  | {:error, Tesla.Env.t()}

Get a customer service request Get a specific customer service request.

Parameters

connection (TelnyxAPI.Connection): Connection to server

customer_service_request_id (String.t): The ID of the customer service request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

verify_phone_number_coverage(connection, u_nknownbasetype, opts \\ [])

@spec verify_phone_number_coverage(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, Map.t()}
  | {:ok, TelnyxAPI.Model.ListCustomerServiceRequestPhoneNumberCoverage.t()}
  | {:error, Tesla.Env.t()}

Verify CSR phone number coverage Verify the coverage for a list of phone numbers.

Parameters

connection (TelnyxAPI.Connection): Connection to server

u_nknownbasetype (UNKNOWN_BASE_TYPE):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure