TelnyxAPI.Api.Documents (telnyx_api v2.0.0)

API calls for all endpoints tagged Documents.

Functions

create_document(connection, create_doc_service_document_request, opts \\ [])

@spec create_document(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateDocServiceDocumentRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.DocServiceDocumentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Upload a document Upload a document.<br /><br />Uploaded files must be linked to a service within 30 minutes or they will be automatically deleted.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_doc_service_document_request (CreateDocServiceDocumentRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_document(connection, id, opts \\ [])

@spec delete_document(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.DocServiceDocumentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Delete a document Delete a document.<br /><br />A document can only be deleted if it's not linked to a service. If it is linked to a service, it must be unlinked prior to deleting.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

download_doc_service_document(connection, id, opts \\ [])

@spec download_doc_service_document(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, String.t()}
  | {:error, Tesla.Env.t()}

Download a document Download a document.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_document_links(connection, opts \\ [])

@spec list_document_links(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListDocServiceDocumentLinksResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List all document links List all documents links ordered by created_at descending.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketdocument_id_right_square_bracket (String.t): Identifies the associated document to filter on.

:filter_left_square_bracketlinked_record_type_right_square_bracket (String.t): The linked_record_type of the document to filter on.

:filter_left_square_bracketlinked_resource_id_right_square_bracket (String.t): The linked_resource_id of the document to filter on.

Returns

on success {:error, Tesla.Env.t} on failure

list_documents(connection, opts \\ [])

@spec list_documents(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListDocServiceDocumentsResposne.t()}
  | {:error, Tesla.Env.t()}

List all documents List all documents ordered by created_at descending.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketfilename_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Filter by string matching part of filename.

:filter_left_square_bracketcustomer_reference_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter documents by a customer references.

:filter_left_square_bracketcustomer_reference_right_square_bracket_left_square_bracketin_right_square_bracket_left_square_bracket_right_square_bracket (String.t): Filter documents by a list of customer references.

:filter_left_square_bracketcreated_at_right_square_bracket_left_square_bracketgt_right_square_bracket (String.t): Filter by created at greater than provided value.

:filter_left_square_bracketcreated_at_right_square_bracket_left_square_bracketlt_right_square_bracket (String.t): Filter by created at less than provided value.

:sort_left_square_bracket_right_square_bracket (String.t): Specifies the sort order for results. If you want to sort by a field in ascending order, include it as a sort parameter. If you want to sort in descending order, prepend a - in front of the field name.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_document(connection, id, opts \\ [])

@spec retrieve_document(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.DocServiceDocumentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a document Retrieve a document.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_document(connection, id, doc_service_document, opts \\ [])

@spec update_document(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.DocServiceDocument.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.DocServiceDocumentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a document Update a document.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

doc_service_document (DocServiceDocument):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure