TelnyxAPI.Api.DynamicEmergencyEndpoints (telnyx_api v2.0.0)

API calls for all endpoints tagged DynamicEmergencyEndpoints.

Functions

create_dyanmic_emergency_endpoint(connection, dynamic_emergency_endpoint, opts \\ [])

@spec create_dyanmic_emergency_endpoint(
  Tesla.Env.client(),
  TelnyxAPI.Model.DynamicEmergencyEndpoint.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.DynamicEmergencyEndpointResponse.t()}
  | {:error, Tesla.Env.t()}

Create a dynamic emergency endpoint. Creates a dynamic emergency endpoints.

Parameters

connection (TelnyxAPI.Connection): Connection to server

dynamic_emergency_endpoint (DynamicEmergencyEndpoint):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_dynamic_emergency_endpoint(connection, id, opts \\ [])

@spec delete_dynamic_emergency_endpoint(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.DynamicEmergencyEndpointResponse.t()}
  | {:error, Tesla.Env.t()}

Delete a dynamic emergency endpoint Deletes the dynamic emergency endpoint based on the ID provided

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Dynamic Emergency Endpoint id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_dynamic_emergency_endpoint(connection, id, opts \\ [])

@spec get_dynamic_emergency_endpoint(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.DynamicEmergencyEndpointResponse.t()}
  | {:error, Tesla.Env.t()}

Get a dynamic emergency endpoint Returns the dynamic emergency endpoint based on the ID provided

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Dynamic Emergency Endpoint id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_dyanmic_emergency_endpoints(connection, opts \\ [])

@spec list_dyanmic_emergency_endpoints(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListDynamicEmergencyEndpoints.t()}
  | {:error, Tesla.Env.t()}

List dynamic emergency endpoints Returns the dynamic emergency endpoints according to filters

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter by status.

:filter_left_square_bracketcountry_code_right_square_bracket (String.t): Filter by country code.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure