TelnyxAPI.Api.FQDNConnections (telnyx_api v2.0.0)

API calls for all endpoints tagged FQDNConnections.

Functions

create_fqdn_connection(connection, create_fqdn_connection_request, opts \\ [])

@spec create_fqdn_connection(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateFqdnConnectionRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FqdnConnectionResponse.t()}
  | {:error, Tesla.Env.t()}

Create an FQDN connection Creates a FQDN connection.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_fqdn_connection_request (CreateFqdnConnectionRequest): Parameters that can be defined during FQDN connection creation

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_fqdn_connection(connection, id, opts \\ [])

@spec delete_fqdn_connection(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FqdnConnectionResponse.t()}
  | {:error, Tesla.Env.t()}

Delete an FQDN connection Deletes an FQDN connection.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): FQDN Connection ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_fqdn_connections(connection, opts \\ [])

@spec list_fqdn_connections(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListFqdnConnectionsResponse.t()}
  | {:error, Tesla.Env.t()}

List FQDN connections Returns a list of your FQDN connections.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketconnection_name_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): If present, connections with <code>connection_name</code> containing the given value will be returned. Matching is not case-sensitive. Requires at least three characters.

:sort (String.t): Specifies the sort order for results. By default sorting direction is ascending. To have the results sorted in descending order add the <code> -</code> prefix.<br/><br/> That is: <ul> <li> <code>connection_name</code>: sorts the result by the <code>connection_name</code> field in ascending order. </li> <li> <code>-connection_name</code>: sorts the result by the <code>connection_name</code> field in descending order. </li> </ul> <br/> If not given, results are sorted by <code>created_at</code> in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_fqdn_connection(connection, id, opts \\ [])

@spec retrieve_fqdn_connection(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FqdnConnectionResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve an FQDN connection Retrieves the details of an existing FQDN connection.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): FQDN Connection ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_fqdn_connection(connection, id, update_fqdn_connection_request, opts \\ [])

@spec update_fqdn_connection(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateFqdnConnectionRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FqdnConnectionResponse.t()}
  | {:error, Tesla.Env.t()}

Update an FQDN connection Updates settings of an existing FQDN connection.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): FQDN Connection ID

update_fqdn_connection_request (UpdateFqdnConnectionRequest): Parameters that can be updated in a FQDN connection

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure