TelnyxAPI.Api.IPs (telnyx_api v2.0.0)

API calls for all endpoints tagged IPs.

Functions

create_ip(connection, opts \\ [])

@spec create_ip(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.IpResponse.t()} | {:error, Tesla.Env.t()}

Create an Ip Create a new IP object.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (CreateIpRequest):

Returns

on success {:error, Tesla.Env.t} on failure

delete_ip(connection, id, opts \\ [])

@spec delete_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.IpResponse.t()} | {:error, Tesla.Env.t()}

Delete an Ip Delete an IP.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_ips(connection, opts \\ [])

@spec list_ips(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListIpsResponse.t()}
  | {:error, Tesla.Env.t()}

List Ips Get all IPs belonging to the user that match the given filters.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketconnection_id_right_square_bracket (String.t): ID of the IP Connection to which this IP should be attached.

:filter_left_square_bracketip_address_right_square_bracket (String.t): IP adddress represented by this resource.

:filter_left_square_bracketport_right_square_bracket (integer()): Port to use when connecting to this IP.

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_ip(connection, id, opts \\ [])

@spec retrieve_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.IpResponse.t()} | {:error, Tesla.Env.t()}

Retrieve an Ip Return the details regarding a specific IP.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_ip(connection, id, opts \\ [])

@spec update_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.IpResponse.t()} | {:error, Tesla.Env.t()}

Update an Ip Update the details of a specific IP.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

:body (UpdateIpRequest):

Returns

on success {:error, Tesla.Env.t} on failure