TelnyxAPI.Api.MDRUsageReports (telnyx_api v2.0.0)

API calls for all endpoints tagged MDRUsageReports.

Functions

delete_usage_report(connection, id, opts \\ [])

@spec delete_usage_report(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.MdrDeleteUsageReportsResponse.t()}
  | {:error, Tesla.Env.t()}

Delete messaging usage report by id

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_usage_report(connection, id, opts \\ [])

@spec get_usage_report(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.MdrGetUsageReportsByIdResponse.t()}
  | {:error, Tesla.Env.t()}

Fetch a single messaging usage report by id

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_usage_report_sync1(connection, aggregation_type, opts \\ [])

@spec get_usage_report_sync1(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.MdrGetSyncUsageReportResponse.t()}
  | {:error, Tesla.Env.t()}

Generate and fetch messaging usage report synchronously. This endpoint will both generate and fetch the messaging report over a specified time period. No polling is necessary but the response may take up to a couple of minutes.

Parameters

connection (TelnyxAPI.Connection): Connection to server

aggregation_type (String.t):

opts (KeywordList): [optional] Optional parameters

:start_date (DateTime.t):

:end_date (DateTime.t):

:profiles ([String.t]):

Returns

on success {:error, Tesla.Env.t} on failure

get_usage_reports(connection, opts \\ [])

@spec get_usage_reports(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.MdrGetUsageReportsResponse.t()}
  | {:error, Tesla.Env.t()}

Fetch all messaging usage reports. Usage reports are aggregated messaging data for specified time period and breakdown

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): Page number

:page_left_square_bracketsize_right_square_bracket (integer()): Size of the page

Returns

on success {:error, Tesla.Env.t} on failure

submit_usage_report(connection, mdr_post_usage_report_request, opts \\ [])

@spec submit_usage_report(
  Tesla.Env.client(),
  TelnyxAPI.Model.MdrPostUsageReportRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.MdrPostUsageReportsResponse.t()}
  | {:error, Tesla.Env.t()}

Submit request for new new messaging usage report. This endpoint will pull and aggregate messaging data in specified time period.

Parameters

connection (TelnyxAPI.Connection): Connection to server

mdr_post_usage_report_request (MdrPostUsageReportRequest): Mdr usage report data

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure