TelnyxAPI.Api.ManagedAccounts (telnyx_api v2.0.0)

API calls for all endpoints tagged ManagedAccounts.

Functions

create_managed_account(connection, create_managed_account_request, opts \\ [])

@spec create_managed_account(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateManagedAccountRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ManagedAccountResponse.t()}
  | {:ok, TelnyxAPI.Model.UnprocessableManagedAccountEntity.t()}
  | {:error, Tesla.Env.t()}

Create a new managed account. Create a new managed account owned by the authenticated user. You need to be explictly approved by Telnyx in order to become a manager account.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_managed_account_request (CreateManagedAccountRequest): Parameters that define the managed account to be created

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

disable_managed_account(connection, id, opts \\ [])

@spec disable_managed_account(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ManagedAccountResponse.t()}
  | {:ok, TelnyxAPI.Model.UnprocessableManagedAccountEntity.t()}
  | {:error, Tesla.Env.t()}

Disables a managed account Disables a managed account, forbidding it to use Telnyx services, including sending or receiving phone calls and SMS messages. Ongoing phone calls will not be affected. The managed account and its sub-users will still be able to log in via the mission control portal.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Managed Account User ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

enable_managed_account(connection, id, opts \\ [])

@spec enable_managed_account(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ManagedAccountResponse.t()}
  | {:ok, TelnyxAPI.Model.UnprocessableManagedAccountEntity.t()}
  | {:error, Tesla.Env.t()}

Enables a managed account Enables a managed account and its sub-users to use Telnyx services.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Managed Account User ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_managed_accounts(connection, opts \\ [])

@spec list_managed_accounts(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListManagedAccountsResponse.t()}
  | {:error, Tesla.Env.t()}

Lists accounts managed by the current user. Lists the accounts managed by the current user. Users need to be explictly approved by Telnyx in order to become manager accounts.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketemail_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): If present, email containing the given value will be returned. Matching is not case-sensitive. Requires at least three characters.

:filter_left_square_bracketemail_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): If present, only returns results with the <code>email</code> matching exactly the value given.

:sort (String.t): Specifies the sort order for results. By default sorting direction is ascending. To have the results sorted in descending order add the <code> -</code> prefix.<br/><br/> That is: <ul> <li> <code>email</code>: sorts the result by the <code>email</code> field in ascending order. </li> <li> <code>-email</code>: sorts the result by the <code>email</code> field in descending order. </li> </ul> <br/> If not given, results are sorted by <code>created_at</code> in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_managed_account(connection, id, opts \\ [])

@spec retrieve_managed_account(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ManagedAccountResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a managed account Retrieves the details of a single managed account.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Managed Account User ID

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_managed_account(connection, id, update_managed_account_request, opts \\ [])

@spec update_managed_account(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateManagedAccountRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ManagedAccountResponse.t()}
  | {:ok, TelnyxAPI.Model.UnprocessableManagedAccountEntity.t()}
  | {:error, Tesla.Env.t()}

Update a managed account Update a single managed account.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Managed Account User ID

update_managed_account_request (UpdateManagedAccountRequest): Parameters that define the updates to the managed account

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure