TelnyxAPI.Api.MediaStorageAPI (telnyx_api v2.0.0)

API calls for all endpoints tagged MediaStorageAPI.

Functions

create_media_storage(connection, upload_media_request, opts \\ [])

@spec create_media_storage(
  Tesla.Env.client(),
  TelnyxAPI.Model.UploadMediaRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.MediaResourceResponse.t()}
  | {:error, Tesla.Env.t()}

Upload media Upload media file to Telnyx so it can be used with other Telnyx services

Parameters

connection (TelnyxAPI.Connection): Connection to server

upload_media_request (UploadMediaRequest): Upload media request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_media_storage(connection, media_name, opts \\ [])

@spec delete_media_storage(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.Errors.t()} | {:error, Tesla.Env.t()}

Deletes stored media Deletes a stored media file.

Parameters

connection (TelnyxAPI.Connection): Connection to server

media_name (String.t): Uniquely identifies a media resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

download_media_storage(connection, media_name, opts \\ [])

@spec download_media_storage(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, String.t()}
  | {:error, Tesla.Env.t()}

Download stored media Downloads a stored media file.

Parameters

connection (TelnyxAPI.Connection): Connection to server

media_name (String.t): Uniquely identifies a media resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_media_storage(connection, media_name, opts \\ [])

@spec get_media_storage(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.MediaResourceResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve stored media Returns the information about a stored media file.

Parameters

connection (TelnyxAPI.Connection): Connection to server

media_name (String.t): Uniquely identifies a media resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_media_storage(connection, opts \\ [])

@spec list_media_storage(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListOfMediaResourcesResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List uploaded media Returns a list of stored media files.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketcontent_type_right_square_bracket_left_square_bracket_right_square_bracket (String.t): Filters files by given content types

Returns

on success {:error, Tesla.Env.t} on failure

update_media_storage(connection, media_name, update_media_request, opts \\ [])

@spec update_media_storage(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateMediaRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.MediaResourceResponse.t()}
  | {:error, Tesla.Env.t()}

Update stored media Updates a stored media file.

Parameters

connection (TelnyxAPI.Connection): Connection to server

media_name (String.t): Uniquely identifies a media resource.

update_media_request (UpdateMediaRequest): Update media request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure