TelnyxAPI.Api.MessagingHostedNumber (telnyx_api v2.0.0)

API calls for all endpoints tagged MessagingHostedNumber.

Functions

create_messaging_hosted_number_order(connection, opts \\ [])

@spec create_messaging_hosted_number_order(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrieveMessagingHostedNumberOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a messaging hosted number order

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (CreateMessagingHostedNumberOrderRequest): Message payload

Returns

on success {:error, Tesla.Env.t} on failure

delete_messaging_hosted_number(connection, id, opts \\ [])

@spec delete_messaging_hosted_number(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.RetrieveMessagingHostedNumberResponse.t()}
  | {:error, Tesla.Env.t()}

Delete a messaging hosted number

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_messaging_hosted_number_order(connection, opts \\ [])

@spec list_messaging_hosted_number_order(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListMessagingHostedNumberOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List messaging hosted number orders

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_messaging_hosted_number_order(connection, id, opts \\ [])

@spec retrieve_messaging_hosted_number_order(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrieveMessagingHostedNumberOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a messaging hosted number order

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

upload_file_messaging_hosted_number_order(connection, id, opts \\ [])

@spec upload_file_messaging_hosted_number_order(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrieveMessagingHostedNumberOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Upload file required for a messaging hosted number order

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

:loa (String.t): Must be a signed LOA for the numbers in the order in PDF format.

:bill (String.t): Must be the last month's bill with proof of ownership of all of the numbers in the order in PDF format.

Returns

on success {:error, Tesla.Env.t} on failure