TelnyxAPI.Api.MobileApplications (telnyx_api v2.0.0)

API calls for all endpoints tagged MobileApplications.

Functions

add_mobile_application(connection, create_mobile_application_request, opts \\ [])

@spec add_mobile_application(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateMobileApplicationRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.MobileApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Adds a mobile application Creates a new mobile application which can be android or ios based on the type parameter

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_mobile_application_request (CreateMobileApplicationRequest): Mobile application parameters that need to be sent in the request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

add_push_credential(connection, mobile_application_id, u_nknownbasetype, opts \\ [])

@spec add_push_credential(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UNKNOWN_BASE_TYPE.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.PushCredentialResponse.t()}
  | {:error, Tesla.Env.t()}

Adds push credential to a given mobile application Creates a new credential for a given mobile application

Parameters

connection (TelnyxAPI.Connection): Connection to server

mobile_application_id (String.t): The unique identifier of a mobile application

u_nknownbasetype (UNKNOWN_BASE_TYPE): Push credential parameters that need to be sent in the request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_mobile_application_by_id(connection, mobile_application_id, opts \\ [])

@spec delete_mobile_application_by_id(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.Errors.t()} | {:error, Tesla.Env.t()}

Deletes a mobile application Deletes a mobile application based on the given id

Parameters

connection (TelnyxAPI.Connection): Connection to server

mobile_application_id (String.t): The unique identifier of a mobile application

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_all_mobile_applications(connection, opts \\ [])

@spec get_all_mobile_applications(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, [TelnyxAPI.Model.MobileApplicationResponse.t()]}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieves mobile applications Retrieves all mobile applications for a user

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

, ...]} on success {:error, Tesla.Env.t} on failure

get_mobile_application_by_id(connection, mobile_application_id, opts \\ [])

@spec get_mobile_application_by_id(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.MobileApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieves a given mobile application Retrieves a mobile application based on the given id

Parameters

connection (TelnyxAPI.Connection): Connection to server

mobile_application_id (String.t): The unique identifier of a mobile application

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_mobile_application_by_id(connection, mobile_application_id, create_mobile_application_request, opts \\ [])

@spec update_mobile_application_by_id(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.CreateMobileApplicationRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.MobileApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Updates a mobile application Updates a mobile application based on the given id

Parameters

connection (TelnyxAPI.Connection): Connection to server

mobile_application_id (String.t): The unique identifier of a mobile application

create_mobile_application_request (CreateMobileApplicationRequest): Mobile application parameters that need to be sent in the request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure