TelnyxAPI.Api.Notifications (telnyx_api v2.0.0)

API calls for all endpoints tagged Notifications.

Functions

create_notification_channels(connection, opts \\ [])

@spec create_notification_channels(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateNotificationChannel.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a notification channel Create a notification channel.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (NotificationChannel): Add a Notification Channel

Returns

on success {:error, Tesla.Env.t} on failure

create_notification_profile(connection, opts \\ [])

@spec create_notification_profile(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateNotificationProfile.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a notification profile Create a notification profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (NotificationProfile): Add a Notification Profile

Returns

on success {:error, Tesla.Env.t} on failure

create_notification_setting(connection, opts \\ [])

@spec create_notification_setting(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateNotificationSettingAsync.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.CreateNotificationSetting.t()}
  | {:error, Tesla.Env.t()}

Add a Notification Setting Add a notification setting.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (NotificationSetting):

Returns

on success {:error, Tesla.Env.t} on failure

delete_notification_channel(connection, id, opts \\ [])

@spec delete_notification_channel(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.DeleteNotificationChannel.t()}
  | {:error, Tesla.Env.t()}

Delete a notification channel Delete a notification channel.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_notification_profile(connection, id, opts \\ [])

@spec delete_notification_profile(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.DeleteNotificationProfiles.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Delete a notification profile Delete a notification profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_notification_setting(connection, id, opts \\ [])

@spec delete_notification_setting(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.DeleteNotificationSetting.t()}
  | {:ok, TelnyxAPI.Model.DeleteNotificationSettingAsync.t()}
  | {:error, Tesla.Env.t()}

Delete a notification setting Delete a notification setting.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

find_notifications_events(connection, opts \\ [])

@spec find_notifications_events(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListNotificationEvents.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List all Notifications Events Returns a list of your notifications events.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

find_notifications_events_conditions(connection, opts \\ [])

@spec find_notifications_events_conditions(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListNotificationEvenConditions.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List all Notifications Events Conditions Returns a list of your notifications events conditions.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketassociated_record_type_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by the associated record type

Returns

on success {:error, Tesla.Env.t} on failure

find_notifications_profiles(connection, opts \\ [])

@spec find_notifications_profiles(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListNotificationProfiles.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List all Notifications Profiles Returns a list of your notifications profiles.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_notification_channels(connection, opts \\ [])

@spec list_notification_channels(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListNotificationChannels.t()}
  | {:error, Tesla.Env.t()}

List notification channels List notification channels.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketchannel_type_id_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by the id of a channel type

Returns

on success {:error, Tesla.Env.t} on failure

list_notification_settings(connection, opts \\ [])

@spec list_notification_settings(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListNotificationSettings.t()}
  | {:error, Tesla.Env.t()}

List notification settings List notification settings.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketnotification_profile_id_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by the id of a notification profile

:filter_left_square_bracketnotification_channel_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by the id of a notification channel

:filter_left_square_bracketnotification_event_condition_id_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by the id of a notification channel

:filter_left_square_bracketassociated_record_type_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by the associated record type

:filter_left_square_bracketstatus_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): The status of a notification setting

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_notification_channel(connection, id, opts \\ [])

@spec retrieve_notification_channel(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.RetrieveNotificationChannel.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a notification channel Retrieve a notification channel.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_notification_profile(connection, id, opts \\ [])

@spec retrieve_notification_profile(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.RetrieveNotificationProfile.t()}
  | {:error, Tesla.Env.t()}

Retrieve a notification profile Retrieve a notification profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_notification_setting(connection, id, opts \\ [])

@spec retrieve_notification_setting(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.RetrieveNotificationSetting.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a notification setting Retrieve a notification setting.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_notification_channel(connection, id, notification_channel, opts \\ [])

@spec update_notification_channel(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.NotificationChannel.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.UpdateNotificationChannel.t()}
  | {:error, Tesla.Env.t()}

Update a notification channel Update a notification channel.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

notification_channel (NotificationChannel): Update notification channel object

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_notification_profile(connection, id, notification_profile, opts \\ [])

@spec update_notification_profile(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.NotificationProfile.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.UpdateNotificationProfile.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a notification profile Update a notification profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

notification_profile (NotificationProfile): Update notification profile object

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure