TelnyxAPI.Api.NumberBlocksBackgroundJobs (telnyx_api v2.0.0)

API calls for all endpoints tagged NumberBlocksBackgroundJobs.

Functions

create_phone_number_blocks_job_delete_phone_number_block(connection, phone_number_blocks_job_delete_phone_number_block_request, opts \\ [])

@spec create_phone_number_blocks_job_delete_phone_number_block(
  Tesla.Env.client(),
  TelnyxAPI.Model.PhoneNumberBlocksJobDeletePhoneNumberBlockRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.PhoneNumberBlocksJobDeletePhoneNumberBlock.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Deletes all numbers associated with a phone number block Creates a new background job to delete all the phone numbers associated with the given block. We will only consider the phone number block as deleted after all phone numbers associated with it are removed, so multiple executions of this job may be necessary in case some of the phone numbers present errors during the deletion process.

Parameters

connection (TelnyxAPI.Connection): Connection to server

phone_number_blocks_job_delete_phone_number_block_request (PhoneNumberBlocksJobDeletePhoneNumberBlockRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_phone_number_blocks_jobs(connection, opts \\ [])

@spec list_phone_number_blocks_jobs(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListPhoneNumberBlocksBackgroundJobsResponse.t()}
  | {:error, Tesla.Env.t()}

Lists the phone number blocks jobs

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_brackettype_right_square_bracket (String.t): Filter the phone number blocks jobs by type.

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter the phone number blocks jobs by status.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:sort (String.t): Specifies the sort order for results. If not given, results are sorted by created_at in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_phone_number_blocks_job(connection, id, opts \\ [])

@spec retrieve_phone_number_blocks_job(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.RetrievePhoneNumberBlocksJob.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieves a phone number blocks job

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the Phone Number Blocks Job.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure