TelnyxAPI.Api.NumberConfigurations (telnyx_api v2.0.0)

API calls for all endpoints tagged NumberConfigurations.

Functions

delete_phone_number(connection, id, opts \\ [])

@spec delete_phone_number(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.PhoneNumberResponse1.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Delete a phone number

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

enable_emergency_phone_number(connection, id, phone_number_enable_emergency_request, opts \\ [])

@spec enable_emergency_phone_number(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.PhoneNumberEnableEmergencyRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreatePhoneNumberEmergencyRequested.t()}
  | {:ok, TelnyxAPI.Model.CreatePhoneNumberEmergencyEnable.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Enable emergency for a phone number

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

phone_number_enable_emergency_request (PhoneNumberEnableEmergencyRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_phone_numbers(connection, opts \\ [])

@spec list_phone_numbers(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListPhoneNumbersResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List phone numbers

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_brackettag_right_square_bracket (String.t): Filter by phone number tags.

:filter_left_square_bracketphone_number_right_square_bracket (String.t): Filter by phone number. Requires at least three digits. Non-numerical characters will result in no values being returned.

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter by phone number status.

:filter_left_square_bracketconnection_id_right_square_bracket (String.t): Filter by connection_id.

:filter_left_square_bracketvoice_periodconnection_name_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Filter contains connection name. Requires at least three characters.

:filter_left_square_bracketvoice_periodconnection_name_right_square_bracket_left_square_bracketstarts_with_right_square_bracket (String.t): Filter starts with connection name. Requires at least three characters.

:filter_left_square_bracketvoice_periodconnection_name_right_square_bracket_left_square_bracketends_with_right_square_bracket (String.t): Filter ends with connection name. Requires at least three characters.

:filter_left_square_bracketvoice_periodconnection_name_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter by connection name.

:filter_left_square_bracketusage_payment_method_right_square_bracket (String.t): Filter by usage_payment_method.

:filter_left_square_bracketbilling_group_id_right_square_bracket (String.t): Filter by the billing_group_id associated with phone numbers. To filter to only phone numbers that have no billing group associated them, set the value of this filter to the string 'null'.

:filter_left_square_bracketemergency_address_id_right_square_bracket (String.t): Filter by the emergency_address_id associated with phone numbers. To filter only phone numbers that have no emergency address associated with them, set the value of this filter to the string 'null'.

:filter_left_square_bracketcustomer_reference_right_square_bracket (String.t): Filter numbers via the customer_reference set.

:sort (String.t): Specifies the sort order for results. If not given, results are sorted by created_at in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

list_phone_numbers_with_messaging_settings(connection, opts \\ [])

@spec list_phone_numbers_with_messaging_settings(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListMessagingSettingsResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List phone numbers with messaging settings

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_phone_numbers_with_voice_settings(connection, opts \\ [])

@spec list_phone_numbers_with_voice_settings(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListPhoneNumbersWithVoiceSettingsResponse.t()}
  | {:error, Tesla.Env.t()}

List phone numbers with voice settings

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketphone_number_right_square_bracket (String.t): Filter by phone number. Requires at least three digits. Non-numerical characters will result in no values being returned.

:filter_left_square_bracketconnection_name_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Filter contains connection name. Requires at least three characters.

:filter_left_square_bracketcustomer_reference_right_square_bracket (String.t): Filter numbers via the customer_reference set.

:filter_left_square_bracketusage_payment_method_right_square_bracket (String.t): Filter by usage_payment_method.

:sort (String.t): Specifies the sort order for results. If not given, results are sorted by created_at in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_phone_number(connection, id, opts \\ [])

@spec retrieve_phone_number(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.PhoneNumberResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a phone number

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_phone_number_with_messaging_settings(connection, id, opts \\ [])

@spec retrieve_phone_number_with_messaging_settings(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrieveMessagingSettingsResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a phone number with messaging settings

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_phone_number_with_voice_settings(connection, id, opts \\ [])

@spec retrieve_phone_number_with_voice_settings(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrievePhoneNumberVoiceResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a phone number with voice settings

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_phone_number(connection, id, update_phone_number_request, opts \\ [])

@spec update_phone_number(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdatePhoneNumberRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.PhoneNumberResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a phone number

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

update_phone_number_request (UpdatePhoneNumberRequest): Updated settings for the phone number.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_phone_number_with_messaging_settings(connection, id, update_phone_number_messaging_settings_request, opts \\ [])

@spec update_phone_number_with_messaging_settings(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdatePhoneNumberMessagingSettingsRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrieveMessagingSettingsResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a phone number with messaging settings

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the type of resource.

update_phone_number_messaging_settings_request (UpdatePhoneNumberMessagingSettingsRequest): Updated messaging settings for the phone number

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_phone_number_with_voice_settings(connection, id, update_phone_number_voice_settings_request, opts \\ [])

@spec update_phone_number_with_voice_settings(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdatePhoneNumberVoiceSettingsRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RetrievePhoneNumberVoiceResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a phone number with voice settings

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

update_phone_number_voice_settings_request (UpdatePhoneNumberVoiceSettingsRequest): Updated voice settings for the phone number.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure