TelnyxAPI.Api.NumberOrders (telnyx_api v2.0.0)

API calls for all endpoints tagged NumberOrders.

Functions

create_comment(connection, comment, opts \\ [])

@spec create_comment(Tesla.Env.client(), TelnyxAPI.Model.Comment.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.CommentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a comment

Parameters

connection (TelnyxAPI.Connection): Connection to server

comment (Comment):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_number_order(connection, create_number_order_request, opts \\ [])

@spec create_number_order(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateNumberOrderRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.NumberOrderResponse.t()}
  | {:error, Tesla.Env.t()}

Create a number order Creates a phone number order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_number_order_request (CreateNumberOrderRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_comments(connection, filter_left_square_bracketcomment_record_type_right_square_bracket, filter_left_square_bracketcomment_record_id_right_square_bracket, opts \\ [])

@spec list_comments(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ListCommentsResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve all comments

Parameters

connection (TelnyxAPI.Connection): Connection to server

filter_left_square_bracketcomment_record_type_right_square_bracket (String.t): Record type that the comment relates to i.e number_order, sub_number_order or number_order_phone_number

filter_left_square_bracketcomment_record_id_right_square_bracket (String.t): ID of the record the comments relate to

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_number_orders(connection, opts \\ [])

@spec list_number_orders(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListNumberOrdersResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List number orders Get a paginated list of number orders.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter number orders by status.

:filter_left_square_bracketcreated_at_right_square_bracket_left_square_bracketgt_right_square_bracket (String.t): Filter number orders later than this value.

:filter_left_square_bracketcreated_at_right_square_bracket_left_square_bracketlt_right_square_bracket (String.t): Filter number orders earlier than this value.

:filter_left_square_bracketphone_numbers_count_right_square_bracket (String.t): Filter number order with this amount of numbers

:filter_left_square_bracketcustomer_reference_right_square_bracket (String.t): Filter number orders via the customer reference set.

:filter_left_square_bracketrequirements_met_right_square_bracket (boolean()): Filter number orders by requirements met.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_regulartory_requirements(connection, filter_left_square_bracketphone_number_right_square_bracket, opts \\ [])

@spec list_regulartory_requirements(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ListRegulatoryRequirements.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve regulartory requirements

Parameters

connection (TelnyxAPI.Connection): Connection to server

filter_left_square_bracketphone_number_right_square_bracket (String.t): Record type phone number/ phone numbers

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_sub_number_orders(connection, opts \\ [])

@spec list_sub_number_orders(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListSubNumberOrdersResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List sub number orders Get a paginated list of sub number orders.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketuser_id_right_square_bracket (String.t): User ID of the user who owns the sub number order

:filter_left_square_bracketorder_request_id_right_square_bracket (String.t): ID of the number order the sub number order belongs to

:filter_left_square_bracketcountry_code_right_square_bracket (String.t): ISO alpha-2 country code.

:filter_left_square_bracketphone_number_type_right_square_bracket (String.t): Phone Number Type

:filter_left_square_bracketphone_numbers_count_right_square_bracket (integer()): Amount of numbers in the sub number order

Returns

on success {:error, Tesla.Env.t} on failure

mark_comment_read(connection, id, opts \\ [])

@spec mark_comment_read(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ReadCommentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Mark a comment as read

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): The comment ID.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_comment(connection, id, opts \\ [])

@spec retrieve_comment(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.CommentResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a comment

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): The comment ID.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_number_order(connection, number_order_id, opts \\ [])

@spec retrieve_number_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.NumberOrderResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a number order Get an existing phone number order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

number_order_id (String.t): The number order ID.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_number_order_phone_number(connection, number_order_phone_number_id, opts \\ [])

@spec retrieve_number_order_phone_number(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.NumberOrderPhoneNumberResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a number order phone number. Get an existing number order phone number.

Parameters

connection (TelnyxAPI.Connection): Connection to server

number_order_phone_number_id (String.t): The number order phone number ID.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_number_order_phone_numbers(connection, opts \\ [])

@spec retrieve_number_order_phone_numbers(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListNumberOrderPhoneNumbersResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a list of phone numbers associated to orders Get a list of phone numbers associated to orders.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_sub_number_order(connection, sub_number_order_id, opts \\ [])

@spec retrieve_sub_number_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SubNumberOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a sub number order Get an existing sub number order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sub_number_order_id (String.t): The sub number order ID.

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketinclude_phone_numbers_right_square_bracket (boolean()): Include the first 50 phone number objects in the results

Returns

on success {:error, Tesla.Env.t} on failure

update_number_order(connection, number_order_id, update_number_order_request, opts \\ [])

@spec update_number_order(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateNumberOrderRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.NumberOrderResponse.t()}
  | {:error, Tesla.Env.t()}

Update a number order Updates a phone number order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

number_order_id (String.t): The number order ID.

update_number_order_request (UpdateNumberOrderRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_number_order_phone_number(connection, number_order_phone_number_id, update_number_order_phone_number_request, opts \\ [])

@spec update_number_order_phone_number(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateNumberOrderPhoneNumberRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.NumberOrderPhoneNumberResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a number order phone number. Updates a number order phone number.

Parameters

connection (TelnyxAPI.Connection): Connection to server

number_order_phone_number_id (String.t): The number order phone number ID.

update_number_order_phone_number_request (UpdateNumberOrderPhoneNumberRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_sub_number_order(connection, sub_number_order_id, update_sub_number_order_request, opts \\ [])

@spec update_sub_number_order(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateSubNumberOrderRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.SubNumberOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a sub number order Updates a sub number order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sub_number_order_id (String.t): The sub number order ID.

update_sub_number_order_request (UpdateSubNumberOrderRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure