TelnyxAPI.Api.NumberPortout (telnyx_api v2.0.0)

API calls for all endpoints tagged NumberPortout.

Functions

find_portout_comments(connection, id, opts \\ [])

@spec find_portout_comments(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortoutComments.t()}
  | {:error, Tesla.Env.t()}

List all comments for a portout request Returns a list of comments for a portout request.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Portout id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

find_portout_request(connection, id, opts \\ [])

@spec find_portout_request(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.PortoutResponse.t()}
  | {:error, Tesla.Env.t()}

Get a portout request Returns the portout request based on the ID provided

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Portout id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_port_request_supporting_documents(connection, id, opts \\ [])

@spec get_port_request_supporting_documents(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.PortOutListSupportingDocumentsResponse.t()}
  | {:error, Tesla.Env.t()}

List supporting documents on a portout request List every supporting documents for a portout request.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Portout id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_portout_request(connection, opts \\ [])

@spec list_portout_request(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortoutResponse.t()}
  | {:error, Tesla.Env.t()}

List portout requests Returns the portout requests according to filters

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketcarrier_name_right_square_bracket (String.t): Filter by new carrier name.

:filter_left_square_bracketspid_right_square_bracket (String.t): Filter by new carrier spid.

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter by portout status.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

post_port_request_comment(connection, id, create_comment_portout_request, opts \\ [])

@spec post_port_request_comment(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.CreateCommentPortoutRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.PortoutCommentResposne.t()}
  | {:error, Tesla.Env.t()}

Create a comment on a portout request Creates a comment on a portout request.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Portout id

create_comment_portout_request (CreateCommentPortoutRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

post_port_request_supporting_documents(connection, id, create_porting_supporting_docs_request, opts \\ [])

@spec post_port_request_supporting_documents(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.CreatePortingSupportingDocsRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.CreatePortOutSupportingDocumentsResponse.t()}
  | {:error, Tesla.Env.t()}

Create a list of supporting documents on a portout request Creates a list of supporting documents on a portout request.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Portout id

create_porting_supporting_docs_request (CreatePortingSupportingDocsRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_portout_request(connection, id, status, update_portout_status_request, opts \\ [])

@spec update_portout_request(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  TelnyxAPI.Model.UpdatePortoutStatusRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.PortoutResponse.t()}
  | {:error, Tesla.Env.t()}

Update Status Authorize or reject portout request

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Portout id

status (String.t): Updated portout status

update_portout_status_request (UpdatePortoutStatusRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure