TelnyxAPI.Api.NumberSearch (telnyx_api v2.0.0)

API calls for all endpoints tagged NumberSearch.

Functions

list_available_phone_number_blocks(connection, opts \\ [])

@spec list_available_phone_number_blocks(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.ListAvailablePhoneNumbersBlocksResponse.t()}
  | {:error, Tesla.Env.t()}

List available phone number blocks

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketstarts_with_right_square_bracket (String.t): Filter number blocks that start with a pattern (meant to be used after national_destination_code filter has been set).

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketends_with_right_square_bracket (String.t): Filter numbers ending with a pattern.

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Filter numbers containing a pattern.

:filter_left_square_bracketlocality_right_square_bracket (String.t): Filter phone numbers by city.

:filter_left_square_bracketadministrative_area_right_square_bracket (String.t): Filter phone numbers by US state/CA province.

:filter_left_square_bracketcountry_code_right_square_bracket (String.t): Filter phone numbers by ISO alpha-2 country code.

:filter_left_square_bracketnational_destination_code_right_square_bracket (String.t): Filter by the national destination code of the number. This filter is only applicable to North American numbers.

:filter_left_square_bracketrate_center_right_square_bracket (String.t): Filter phone numbers by NANP rate center. This filter is only applicable to North American numbers.

:filter_left_square_bracketnumber_type_right_square_bracket (String.t): Filter phone numbers by number type.

:filter_left_square_bracketfeatures_right_square_bracket ([String.t]): Filter if the phone number should be used for voice, fax, mms, sms, emergency.

:filter_left_square_bracketminimum_block_size_right_square_bracket (integer()): Filter number blocks by minimum blocks size

:filter_left_square_bracketlimit_right_square_bracket (integer()): Limits the number of results.

Returns

on success {:error, Tesla.Env.t} on failure

list_available_phone_numbers(connection, opts \\ [])

@spec list_available_phone_numbers(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListAvailablePhoneNumbersResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List available phone numbers

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketstarts_with_right_square_bracket (String.t): Filter numbers starting with a pattern (exclude NDC from start of this filter if used with national_destination_code filter).

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketends_with_right_square_bracket (String.t): Filter numbers ending with a pattern (max length of 4 digits if used with national_destination_code filter).

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Filter numbers containing a pattern (must be used with national_destination_code filter and only searches within last 4 digits).

:filter_left_square_bracketlocality_right_square_bracket (String.t): Filter phone numbers by city.

:filter_left_square_bracketadministrative_area_right_square_bracket (String.t): Filter phone numbers by US state/CA province.

:filter_left_square_bracketcountry_code_right_square_bracket (String.t): Filter phone numbers by ISO alpha-2 country code.

:filter_left_square_bracketnational_destination_code_right_square_bracket (String.t): Filter by the national destination code of the number. This filter is only applicable to North American numbers.

:filter_left_square_bracketrate_center_right_square_bracket (String.t): Filter phone numbers by NANP rate center. This filter is only applicable to North American numbers.

:filter_left_square_bracketphone_number_type_right_square_bracket (String.t): Filter phone numbers by number type.

:filter_left_square_bracketfeatures_right_square_bracket ([String.t]): Filter if the phone number should be used for voice, fax, mms, sms, emergency.

:filter_left_square_bracketlimit_right_square_bracket (integer()): Limits the number of results.

:filter_left_square_bracketbest_effort_right_square_bracket (boolean()): Filter to determine if best effort results should be included. Only available in NANPA countries.

:filter_left_square_bracketquickship_right_square_bracket (boolean()): Filter to exclude phone numbers that need additional time after to purchase to receive phone calls. Only available for toll-free numbers.

:filter_left_square_bracketreservable_right_square_bracket (boolean()): Filter to exclude phone numbers that cannot be reserved before purchase.

:filter_left_square_bracketexclude_held_numbers_right_square_bracket (boolean()): Filter to exclude phone numbers that are currently on hold for your account.

Returns

on success {:error, Tesla.Env.t} on failure