TelnyxAPI.Api.OutboundVoiceProfiles (telnyx_api v2.0.0)

API calls for all endpoints tagged OutboundVoiceProfiles.

Functions

create_outbound_voice_profile(connection, create_outbound_voice_profile_request, opts \\ [])

@spec create_outbound_voice_profile(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateOutboundVoiceProfileRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.OutboundVoiceProfileResponse.t()}
  | {:error, Tesla.Env.t()}

Create an outbound voice profile Create an outbound voice profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_outbound_voice_profile_request (CreateOutboundVoiceProfileRequest): Parameters that can be defined when creating an outbound voice profile

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_outbound_voice_profile(connection, id, opts \\ [])

@spec delete_outbound_voice_profile(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.OutboundVoiceProfileResponse.t()}
  | {:error, Tesla.Env.t()}

Delete an outbound voice profile Deletes an existing outbound voice profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_outbound_voice_profiles(connection, opts \\ [])

@spec list_outbound_voice_profiles(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListOutboundVoiceProfilesResponse.t()}
  | {:error, Tesla.Env.t()}

Get all outbound voice profiles Get all outbound voice profiles belonging to the user that match the given filters.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketname_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Optional filter on outbound voice profile name.

:sort (String.t): Specifies the sort order for results. By default sorting direction is ascending. To have the results sorted in descending order add the <code>-</code> prefix.<br/><br/> That is: <ul> <li> <code>name</code>: sorts the result by the <code>name</code> field in ascending order. </li> <li> <code>-name</code>: sorts the result by the <code>name</code> field in descending order. </li> </ul> <br/>

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_outbound_voice_profile(connection, id, opts \\ [])

@spec retrieve_outbound_voice_profile(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.OutboundVoiceProfileResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve an outbound voice profile Retrieves the details of an existing outbound voice profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_outbound_voice_profile(connection, id, update_outbound_voice_profile_request, opts \\ [])

@spec update_outbound_voice_profile(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateOutboundVoiceProfileRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.OutboundVoiceProfileResponse.t()}
  | {:error, Tesla.Env.t()}

Updates an existing outbound voice profile. Updates an existing outbound voice profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

update_outbound_voice_profile_request (UpdateOutboundVoiceProfileRequest): Parameters that can be updated on an outbound voice profile

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure