TelnyxAPI.Api.PortingOrder (telnyx_api v2.0.0)

API calls for all endpoints tagged PortingOrder.

Functions

activate_porting_order(connection, id, opts \\ [])

@spec activate_porting_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrdersActivationJob.t()}
  | {:error, Tesla.Env.t()}

Activates every number on a porting order. Activate every numbers of a porting order asynchronously.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

cancel_porting_order(connection, id, opts \\ [])

@spec cancel_porting_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrder.t()}
  | {:error, Tesla.Env.t()}

Cancel a porting order Cancel a porting order

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

confirm_porting_order(connection, id, opts \\ [])

@spec confirm_porting_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrder.t()}
  | {:error, Tesla.Env.t()}

Submit a porting order. Confirm and submit your porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_porting_order(connection, create_porting_order, opts \\ [])

@spec create_porting_order(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreatePortingOrder.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListDraftPortingOrderWithoutPagination.t()}
  | {:error, Tesla.Env.t()}

Create a porting order Creates a new porting order object.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_porting_order (CreatePortingOrder):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_porting_order_comment(connection, id, create_porting_order_comment, opts \\ [])

@spec create_porting_order_comment(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.CreatePortingOrderComment.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrderComment.t()}
  | {:error, Tesla.Env.t()}

Create a comment for a porting order Creates a new comment for a porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

create_porting_order_comment (CreatePortingOrderComment):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_porting_order(connection, id, opts \\ [])

@spec delete_porting_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:error, Tesla.Env.t()}

Delete a porting order Deletes an existing porting order. This operation is restrict to porting orders in draft state.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_porting_order(connection, id, opts \\ [])

@spec get_porting_order(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrder.t()}
  | {:error, Tesla.Env.t()}

Retrieve a porting order Retrieves the details of an existing porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

:include_phone_numbers (boolean()): Include the first 50 phone number objects in the results

Returns

on success {:error, Tesla.Env.t} on failure

get_porting_order_loa_template(connection, id, opts \\ [])

@spec get_porting_order_loa_template(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, String.t()} | {:error, Tesla.Env.t()}

Download a porting order loa template Download a porting order loa template

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_porting_order_sub_request(connection, id, opts \\ [])

@spec get_porting_order_sub_request(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.SubRequestByPortingOrder.t()}
  | {:error, Tesla.Env.t()}

Retrieve the associated V1 sub_request_id and port_request_id Retrieve the associated V1 sub_request_id and port_request_id

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_porting_orders_activation_job(connection, id, activation_job_id, opts \\ [])

@spec get_porting_orders_activation_job(
  Tesla.Env.client(),
  String.t(),
  String.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrdersActivationJob.t()}
  | {:error, Tesla.Env.t()}

Retrieve a porting activation job Returns a porting activation job.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

activation_job_id (String.t): Activation Job Identifier

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_allowed_foc_windows(connection, id, opts \\ [])

@spec list_allowed_foc_windows(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListAllowedFocWindow.t()}
  | {:error, Tesla.Env.t()}

List allowed FOC dates Returns a list of allowed FOC dates for a porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_porting_order_documents(connection, id, opts \\ [])

@spec list_porting_order_documents(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortingOrderDocuments.t()}
  | {:error, Tesla.Env.t()}

List all documents associated with a porting order Returns a list of all documents associated with a porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_porting_orders(connection, opts \\ [])

@spec list_porting_orders(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortingOrder.t()}
  | {:error, Tesla.Env.t()}

List all porting orders Returns a list of your porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:include_phone_numbers (boolean()): Include the first 50 phone number objects in the results

:filter_left_square_bracketstatus_right_square_bracket (String.t): Filter results by status

:filter_left_square_bracketcustomer_reference_right_square_bracket (String.t): Filter results by user reference

:filter_left_square_bracketphone_numbers_right_square_bracket_left_square_bracketcountry_code_right_square_bracket (String.t): Filter results by country ISO 3166-1 alpha-2 code

:filter_left_square_bracketphone_numbers_right_square_bracket_left_square_bracketcarrier_name_right_square_bracket (String.t): Filter results by old service provider

:filter_left_square_bracketmisc_right_square_bracket_left_square_brackettype_right_square_bracket (TelnyxAPI.Model.PortingOrderType.t): Filter results by porting order type

:filter_left_square_bracketend_user_right_square_bracket_left_square_bracketadmin_right_square_bracket_left_square_bracketentity_name_right_square_bracket (String.t): Filter results by person or company name

:filter_left_square_bracketend_user_right_square_bracket_left_square_bracketadmin_right_square_bracket_left_square_bracketauth_person_name_right_square_bracket (String.t): Filter results by authorized person

:filter_left_square_bracketactivation_settings_right_square_bracket_left_square_bracketfast_port_eligible_right_square_bracket (boolean()): Filter results by fast port eligible

:filter_left_square_bracketactivation_settings_right_square_bracket_left_square_bracketfoc_datetime_requested_right_square_bracket_left_square_bracketgt_right_square_bracket (String.t): Filter results by foc date later than this value

:filter_left_square_bracketactivation_settings_right_square_bracket_left_square_bracketfoc_datetime_requested_right_square_bracket_left_square_bracketlt_right_square_bracket (String.t): Filter results by foc date earlier than this value

:sort (String.t): Specifies the sort order for results. If not given, results are sorted by created_at in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

list_porting_orders_activation_jobs(connection, id, opts \\ [])

@spec list_porting_orders_activation_jobs(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortingOrdersActivationJobs.t()}
  | {:error, Tesla.Env.t()}

List all porting activation jobs Returns a list of your porting activation jobs.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_porting_orders_comments(connection, id, opts \\ [])

@spec list_porting_orders_comments(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortingOrdersComments.t()}
  | {:error, Tesla.Env.t()}

List all comments of a porting order Returns a list of all comments of a porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_porting_orders_exception_types(connection, opts \\ [])

@spec list_porting_orders_exception_types(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortingOrdersExceptionTypes.t()}
  | {:error, Tesla.Env.t()}

List all exception types Returns a list of all possible exception types for a porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_porting_phone_numbers(connection, opts \\ [])

@spec list_porting_phone_numbers(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListPortingPhoneNumbers.t()}
  | {:error, Tesla.Env.t()}

List all porting phone numbers Returns a list of your porting phone numbers.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketporting_order_id_right_square_bracket (String.t): Filter results by porting order id

:filter_left_square_bracketporting_order_id_right_square_bracket_left_square_bracketin_right_square_bracket_left_square_bracket_right_square_bracket ([String.t]): Filter results by a list of porting order ids

:filter_left_square_bracketsupport_key_right_square_bracket_left_square_bracketeq_right_square_bracket (String.t): Filter results by support key

:filter_left_square_bracketsupport_key_right_square_bracket_left_square_bracketin_right_square_bracket_left_square_bracket_right_square_bracket ([String.t]): Filter results by a list of support keys

:filter_left_square_bracketphone_number_right_square_bracket (String.t): Filter results by phone number

:filter_left_square_bracketphone_number_right_square_bracket_left_square_bracketin_right_square_bracket_left_square_bracket_right_square_bracket ([String.t]): Filter results by a list of phone numbers

:filter_left_square_bracketporting_order_status_right_square_bracket (String.t): Filter results by porting order status

:filter_left_square_bracketactivation_status_right_square_bracket (TelnyxAPI.Model.PortingOrderActivationStatus.t): Filter results by activation status

:filter_left_square_bracketportability_status_right_square_bracket (TelnyxAPI.Model.PortabilityStatus.t): Filter results by portability status

Returns

on success {:error, Tesla.Env.t} on failure

update_porting_order(connection, id, update_porting_order, opts \\ [])

@spec update_porting_order(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdatePortingOrder.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ShowPortingOrder.t()}
  | {:error, Tesla.Env.t()}

Edit a porting order Edits the details of an existing porting order. Any or all of a porting order’s attributes may be included in the resource object included in a PATCH request. If a request does not include all of the attributes for a resource, the system will interpret the missing attributes as if they were included with their current values. To explicitly set something to null, it must be included in the request with a null value.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

update_porting_order (UpdatePortingOrder):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

upload_porting_order_documents(connection, id, u_nknownbasetype, opts \\ [])

Associate a list of documents with a porting order Associate a list of documents with a porting order.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Porting Order id

u_nknownbasetype (UNKNOWN_BASE_TYPE):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure