TelnyxAPI.Api.PrivateWirelessGateways (telnyx_api v2.0.0)

API calls for all endpoints tagged PrivateWirelessGateways.

Functions

create_private_wireless_gateway(connection, create_private_wireless_gatweway_request, opts \\ [])

@spec create_private_wireless_gateway(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreatePrivateWirelessGatwewayRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.CreatePrivateWirelessGatewayResponse.t()}
  | {:error, Tesla.Env.t()}

Create a Private Wireless Gateway Asynchronously create a Private Wireless Gateway for SIM cards for a previously created network.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_private_wireless_gatweway_request (CreatePrivateWirelessGatwewayRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_private_wireless_gateway(connection, id, opts \\ [])

@spec delete_private_wireless_gateway(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.DeletePrivateWirelessGatewayResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Delete a Private Wireless Gateway Deletes the Private Wireless Gateway.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_private_wireless_gateway(connection, id, opts \\ [])

@spec get_private_wireless_gateway(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetPrivateWirelessGatewayResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Get a Private Wireless Gateway Retrieve information about a Private Wireless Gateway.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_private_wireless_gateways(connection, opts \\ [])

@spec get_private_wireless_gateways(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.GetAllPrivateWirelessGatewaysResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Get all Private Wireless Gateways Get all Private Wireless Gateways belonging to the user.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketname_right_square_bracket (String.t): The name of the Private Wireless Gateway.

:filter_left_square_bracketip_range_right_square_bracket (String.t): The IP address range of the Private Wireless Gateway.

:filter_left_square_bracketregion_code_right_square_bracket (String.t): The name of the region where the Private Wireless Gateway is deployed.

:filter_left_square_bracketcreated_at_right_square_bracket (String.t): Private Wireless Gateway resource creation date.

:filter_left_square_bracketupdated_at_right_square_bracket (String.t): When the Private Wireless Gateway was last updated.

Returns

on success {:error, Tesla.Env.t} on failure