TelnyxAPI.Api.ProgrammableFaxApplications (telnyx_api v2.0.0)

API calls for all endpoints tagged ProgrammableFaxApplications.

Functions

create_fax_application(connection, create_fax_application_request, opts \\ [])

@spec create_fax_application(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateFaxApplicationRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FaxApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Creates a Fax Application Creates a new Fax Application based on the parameters sent in the request. The application name and webhook URL are required. Once created, you can assign phone numbers to your application using the /phone_numbers endpoint.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_fax_application_request (CreateFaxApplicationRequest): Parameters that can be set when creating a Fax Application

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_fax_application(connection, id, opts \\ [])

@spec delete_fax_application(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FaxApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Deletes a Fax Application Permanently deletes a Fax Application. Deletion may be prevented if the application is in use by phone numbers.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_fax_application(connection, id, opts \\ [])

@spec get_fax_application(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FaxApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a Fax Application Return the details of an existing Fax Application inside the 'data' attribute of the response.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_fax_applications(connection, opts \\ [])

@spec list_fax_applications(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.GetAllFaxApplicationsResponse.t()}
  | {:error, Tesla.Env.t()}

List all Fax Applications This endpoint returns a list of your Fax Applications inside the 'data' attribute of the response. You can adjust which applications are listed by using filters. Fax Applications are used to configure how you send and receive faxes using the Programmable Fax API with Telnyx.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketapplication_name_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): If present, applications with <code>application_name</code> containing the given value will be returned. Matching is not case-sensitive. Requires at least three characters.

:filter_left_square_bracketoutbound_periodoutbound_voice_profile_id_right_square_bracket (String.t): Identifies the associated outbound voice profile.

:sort (String.t): Specifies the sort order for results. By default sorting direction is ascending. To have the results sorted in descending order add the <code> -</code> prefix.<br/><br/> That is: <ul> <li> <code>connection_name</code>: sorts the result by the <code>connection_name</code> field in ascending order. </li> <li> <code>-connection_name</code>: sorts the result by the <code>connection_name</code> field in descending order. </li> </ul> <br/> If not given, results are sorted by <code>created_at</code> in descending order.

Returns

on success {:error, Tesla.Env.t} on failure

update_fax_application(connection, id, update_fax_application_request, opts \\ [])

@spec update_fax_application(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateFaxApplicationRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.FaxApplicationResponse.t()}
  | {:error, Tesla.Env.t()}

Update a Fax Application Updates settings of an existing Fax Application based on the parameters of the request.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

update_fax_application_request (UpdateFaxApplicationRequest): Parameters to be updated for the Fax Application

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure