TelnyxAPI.Api.QueueCommands (telnyx_api v0.1.0)

API calls for all endpoints tagged QueueCommands.

Functions

list_queue_calls(connection, queue_name, opts \\ [])

Specs

list_queue_calls(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListQueueCallsResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve calls from a queue Retrieve the list of calls in an existing queue

Parameters

connection (TelnyxAPI.Connection): Connection to server

queue_name (String.t): Uniquely identifies the queue by name

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page
Returns

on success {:error, Tesla.Env.t} on failure

retrieve_call_from_queue(connection, queue_name, call_control_id, opts \\ [])

Specs

retrieve_call_from_queue(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.QueueCallResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a call from a queue Retrieve an existing call from an existing queue

Parameters

connection (TelnyxAPI.Connection): Connection to server

queue_name (String.t): Uniquely identifies the queue by name

call_control_id (String.t): Unique identifier and token for controlling the call

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure

retrieve_call_queue(connection, queue_name, opts \\ [])

Specs

retrieve_call_queue(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.QueueResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a call queue Retrieve an existing call queue

Parameters

connection (TelnyxAPI.Connection): Connection to server

queue_name (String.t): Uniquely identifies the queue by name

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure