TelnyxAPI.Api.RecordingsCommands (telnyx_api v0.1.0)

API calls for all endpoints tagged RecordingsCommands.

Functions

list_recordings(connection, opts \\ [])

Specs

list_recordings(Tesla.Env.client(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListRecordingsResponse.t()}
  | {:error, Tesla.Env.t()}

List recordings Lists recordings for the authenticated user

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketconference_id_right_square_bracket (String.t): Return only recordings associated with a given conference

:filter_left_square_bracketcreated_at_right_square_bracket_left_square_bracketgte_right_square_bracket (String.t): Return only recordings created later than or at given ISO 8601 datetime

:filter_left_square_bracketcreated_at_right_square_bracket_left_square_bracketlte_right_square_bracket (String.t): Return only recordings created earlier than or at given ISO 8601 datetime

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page
Returns

on success {:error, Tesla.Env.t} on failure

retrieve_recording(connection, id, opts \\ [])

Specs

retrieve_recording(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.GetRecordingResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a recording Retrieve a recording from the authenticated user's recordings

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the recording

opts (KeywordList): [optional] Optional parameters
Returns

on success {:error, Tesla.Env.t} on failure