TelnyxAPI.Api.Reporting (telnyx_api v2.0.0)

API calls for all endpoints tagged Reporting.

Functions

create_wdr_report(connection, wdr_report_request, opts \\ [])

@spec create_wdr_report(
  Tesla.Env.client(),
  TelnyxAPI.Model.WdrReportRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateWdrReportResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a Wireless Detail Records (WDRs) Report Asynchronously create a report containing Wireless Detail Records (WDRs) for the SIM cards that consumed wireless data in the given time period.

Parameters

connection (TelnyxAPI.Connection): Connection to server

wdr_report_request (WdrReportRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_wdr_report(connection, id, opts \\ [])

@spec delete_wdr_report(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.DeleteWdrReportResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Delete a Wireless Detail Record (WDR) Report Deletes one specific WDR report

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_wdr_report(connection, id, opts \\ [])

@spec get_wdr_report(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetWdrReportResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Get a Wireless Detail Record (WDR) Report Returns one specific WDR report

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

get_wdr_reports(connection, opts \\ [])

@spec get_wdr_reports(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.GetWdrReportsResponse.t()}
  | {:error, Tesla.Env.t()}

Get all Wireless Detail Records (WDRs) Reports Returns the WDR Reports that match the given parameters.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure