TelnyxAPI.Api.RequirementTypes (telnyx_api v2.0.0)

API calls for all endpoints tagged RequirementTypes.

Functions

doc_reqs_list_requirement_types(connection, opts \\ [])

@spec doc_reqs_list_requirement_types(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.DocReqsListRequirementTypesResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List all requirement types List all requirement types ordered by created_at descending

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketname_right_square_bracket_left_square_bracketcontains_right_square_bracket (String.t): Filters requirement types to those whose name contains a certain string.

:sort_left_square_bracket_right_square_bracket (String.t): Specifies the sort order for results. If you want to sort by a field in ascending order, include it as a sort parameter. If you want to sort in descending order, prepend a - in front of the field name.

Returns

on success {:error, Tesla.Env.t} on failure

doc_reqs_retrieve_requirement_type(connection, id, opts \\ [])

@spec doc_reqs_retrieve_requirement_type(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.DocReqsRequirementTypeResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a requirement types Retrieve a requirement type by id

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the requirement_type record

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure