TelnyxAPI.Api.RoomSessions (telnyx_api v2.0.0)

API calls for all endpoints tagged RoomSessions.

Functions

kick_participant_in_session(connection, room_session_id, actions_participants_request, opts \\ [])

@spec kick_participant_in_session(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ActionsParticipantsRequest.t(),
  keyword()
) :: {:ok, TelnyxAPI.Model.ActionSuccessResponse.t()} | {:error, Tesla.Env.t()}

Kick participants from a room session.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_session_id (String.t): The unique identifier of a room session.

actions_participants_request (ActionsParticipantsRequest): Parameters that can be defined during Kick action.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_room_sessions(connection, opts \\ [])

@spec list_room_sessions(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListRoomSessionsResponse.t()} | {:error, Tesla.Env.t()}

View a list of room sessions.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions created on that date.

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions created after that date.

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions created before that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions updated on that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions updated after that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions updated before that date.

:filter_left_square_bracketdate_ended_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions ended on that date.

:filter_left_square_bracketdate_ended_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions ended after that date.

:filter_left_square_bracketdate_ended_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions ended before that date.

:filter_left_square_bracketroom_id_right_square_bracket (String.t): Room_id for filtering room sessions.

:filter_left_square_bracketactive_right_square_bracket (boolean()): Filter active or inactive room sessions.

:include_participants (boolean()): To decide if room participants should be included in the response.

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

Returns

on success {:error, Tesla.Env.t} on failure

mute_participant_in_session(connection, room_session_id, actions_participants_request, opts \\ [])

@spec mute_participant_in_session(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ActionsParticipantsRequest.t(),
  keyword()
) :: {:ok, TelnyxAPI.Model.ActionSuccessResponse.t()} | {:error, Tesla.Env.t()}

Mute participants in room session.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_session_id (String.t): The unique identifier of a room session.

actions_participants_request (ActionsParticipantsRequest): Parameters that can be defined during Mute action.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

nested_list_room_participants(connection, room_session_id, opts \\ [])

@spec nested_list_room_participants(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ListRoomParticipantsResponse.t()}
  | {:error, Tesla.Env.t()}

View a list of room participants.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_session_id (String.t): The unique identifier of a room session.

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketdate_joined_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room participants that joined on that date.

:filter_left_square_bracketdate_joined_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room participants that joined after that date.

:filter_left_square_bracketdate_joined_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room participants that joined before that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room participants updated on that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room participants updated after that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room participants updated before that date.

:filter_left_square_bracketdate_left_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room participants that left on that date.

:filter_left_square_bracketdate_left_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room participants that left after that date.

:filter_left_square_bracketdate_left_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room participants that left before that date.

:filter_left_square_bracketcontext_right_square_bracket (String.t): Filter room participants based on the context.

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

Returns

on success {:error, Tesla.Env.t} on failure

unmute_participant_in_session(connection, room_session_id, actions_participants_request, opts \\ [])

@spec unmute_participant_in_session(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ActionsParticipantsRequest.t(),
  keyword()
) :: {:ok, TelnyxAPI.Model.ActionSuccessResponse.t()} | {:error, Tesla.Env.t()}

Unmute participants in room session.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_session_id (String.t): The unique identifier of a room session.

actions_participants_request (ActionsParticipantsRequest): Parameters that can be defined during Unmute action.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

view_room_session(connection, room_session_id, opts \\ [])

@spec view_room_session(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetRoomSessionResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

View a room session.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_session_id (String.t): The unique identifier of a room session.

opts (KeywordList): [optional] Optional parameters

:include_participants (boolean()): To decide if room participants should be included in the response.

Returns

on success {:error, Tesla.Env.t} on failure