TelnyxAPI.Api.Rooms (telnyx_api v2.0.0)

API calls for all endpoints tagged Rooms.

Functions

create_room(connection, create_room_request, opts \\ [])

@spec create_room(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateRoomRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateRoomResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a room. Synchronously create a Room.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_room_request (CreateRoomRequest): Parameters that can be defined during room creation.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_room(connection, room_id, opts \\ [])

@spec delete_room(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil} | {:ok, TelnyxAPI.Model.Error.t()} | {:error, Tesla.Env.t()}

Delete a room. Synchronously delete a Room. Participants from that room will be kicked out, they won't be able to join that room anymore, and you won't be charged anymore for that room.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_id (String.t): The unique identifier of a room.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_rooms(connection, opts \\ [])

@spec list_rooms(
  Tesla.Env.client(),
  keyword()
) :: {:ok, TelnyxAPI.Model.ListRoomsResponse.t()} | {:error, Tesla.Env.t()}

View a list of rooms.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering rooms created on that date.

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering rooms created after that date.

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering rooms created before that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering rooms updated on that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering rooms updated after that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering rooms updated before that date.

:filter_left_square_bracketunique_name_right_square_bracket (String.t): Unique_name for filtering rooms.

:include_sessions (boolean()): To decide if room sessions should be included in the response.

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

Returns

on success {:error, Tesla.Env.t} on failure

nested_list_room_sessions(connection, room_id, opts \\ [])

@spec nested_list_room_sessions(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ListRoomSessionsResponse.t()} | {:error, Tesla.Env.t()}

View a list of room sessions.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_id (String.t): The unique identifier of a room.

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions created on that date.

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions created after that date.

:filter_left_square_bracketdate_created_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions created before that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions updated on that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions updated after that date.

:filter_left_square_bracketdate_updated_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions updated before that date.

:filter_left_square_bracketdate_ended_at_right_square_bracket_left_square_bracketeq_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions ended on that date.

:filter_left_square_bracketdate_ended_at_right_square_bracket_left_square_bracketgte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions ended after that date.

:filter_left_square_bracketdate_ended_at_right_square_bracket_left_square_bracketlte_right_square_bracket (Date.t): ISO 8601 date for filtering room sessions ended before that date.

:filter_left_square_bracketactive_right_square_bracket (boolean()): Filter active or inactive room sessions.

:include_participants (boolean()): To decide if room participants should be included in the response.

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

Returns

on success {:error, Tesla.Env.t} on failure

patch_room(connection, room_id, patch_room_request, opts \\ [])

@spec patch_room(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.PatchRoomRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.PatchRoomResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Update a room. Synchronously update a Room.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_id (String.t): The unique identifier of a room.

patch_room_request (PatchRoomRequest): Parameters that can be defined during room update.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

view_room(connection, room_id, opts \\ [])

@spec view_room(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetRoomResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

View a room.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_id (String.t): The unique identifier of a room.

opts (KeywordList): [optional] Optional parameters

:include_sessions (boolean()): To decide if room sessions should be included in the response.

Returns

on success {:error, Tesla.Env.t} on failure