TelnyxAPI.Api.RoomsClientTokens (telnyx_api v2.0.0)

API calls for all endpoints tagged RoomsClientTokens.

Functions

create_room_client_token(connection, room_id, create_room_client_token_request, opts \\ [])

@spec create_room_client_token(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.CreateRoomClientTokenRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateRoomClientTokenResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Create Client Token to join a room. Synchronously create an Client Token to join a Room. Client Token is necessary to join a Telnyx Room. Client Token will expire after token_ttl_secs, a Refresh Token is also provided to refresh a Client Token, the Refresh Token expires after refresh_token_ttl_secs.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_id (String.t): The unique identifier of a room.

create_room_client_token_request (CreateRoomClientTokenRequest): Parameters that can be defined during Room Client Token creation.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

refresh_room_client_token(connection, room_id, refresh_room_client_token_request, opts \\ [])

@spec refresh_room_client_token(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.RefreshRoomClientTokenRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RefreshRoomClientTokenResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Refresh Client Token to join a room. Synchronously refresh an Client Token to join a Room. Client Token is necessary to join a Telnyx Room. Client Token will expire after token_ttl_secs.

Parameters

connection (TelnyxAPI.Connection): Connection to server

room_id (String.t): The unique identifier of a room.

refresh_room_client_token_request (RefreshRoomClientTokenRequest): Parameters that can be defined during Room Client Token refresh.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure