TelnyxAPI.Api.SIMCardOrders (telnyx_api v2.0.0)

API calls for all endpoints tagged SIMCardOrders.

Functions

sim_card_order_get(connection, id, opts \\ [])

@spec sim_card_order_get(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.GetSimCardOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Get a single SIM card order Get a single SIM card order by its ID.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_orders_get(connection, opts \\ [])

@spec sim_card_orders_get(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.GetAllSimCardOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Get all SIM card orders Get all SIM card orders according to filters.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketcreated_at_right_square_bracket (String.t): Filter by ISO 8601 formatted date-time string matching resource creation date-time.

:filter_left_square_bracketupdated_at_right_square_bracket (String.t): Filter by ISO 8601 formatted date-time string matching resource last update date-time.

:filter_left_square_bracketquantity_right_square_bracket (integer()): Filter orders by how many SIM cards were ordered.

:filter_left_square_bracketcost_periodamount_right_square_bracket (String.t): The total monetary amount of the order.

:filter_left_square_bracketcost_periodcurrency_right_square_bracket (String.t): Filter by ISO 4217 currency string.

:filter_left_square_bracketaddress_periodid_right_square_bracket (String.t): Uniquely identifies the address for the order.

:filter_left_square_bracketaddress_periodstreet_address_right_square_bracket (String.t): Returns entries with matching name of the street where the address is located.

:filter_left_square_bracketaddress_periodextended_address_right_square_bracket (String.t): Returns entries with matching name of the supplemental field for address information.

:filter_left_square_bracketaddress_periodlocality_right_square_bracket (String.t): Filter by the name of the city where the address is located.

:filter_left_square_bracketaddress_periodadministrative_area_right_square_bracket (String.t): Filter by state or province where the address is located.

:filter_left_square_bracketaddress_periodcountry_code_right_square_bracket (String.t): Filter by the mobile operator two-character (ISO 3166-1 alpha-2) origin country code.

:filter_left_square_bracketaddress_periodpostal_code_right_square_bracket (String.t): Filter by postal code for the address.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_orders_post(connection, sim_card_order_create, opts \\ [])

@spec sim_card_orders_post(
  Tesla.Env.client(),
  TelnyxAPI.Model.SimCardOrderCreate.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateSimCardOrderResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a SIM card order Creates a new order for SIM cards.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_order_create (SimCardOrderCreate):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_orders_preview(connection, opts \\ [])

@spec sim_card_orders_preview(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.SimCardOrdersPreviewResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Preview SIM card orders Preview SIM card order purchases.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (CreateSimCardOrderPreview):

Returns

on success {:error, Tesla.Env.t} on failure