TelnyxAPI.Api.SIMCards (telnyx_api v2.0.0)

API calls for all endpoints tagged SIMCards.

Functions

bulk_sim_card_network_preferences(connection, opts \\ [])

@spec bulk_sim_card_network_preferences(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.BulkSimCardNetworkPreferenceResponse.t()}
  | {:error, Tesla.Env.t()}

Bulk Network Preferences for SIM cards This API allows dispatching the same operation described for the PUT sim_cards/:sim_card_id/network_preferences API for multiple SIM cards at once.<br/><br/> Although, a SIM card network preference may fail individually under any validation triggered as a consequence of its state. For example, a SIM can't have an in-progress OTA update for applying a Network Preference, so they'll fail when requested in this API. In that scenario, the specific error will be present in the response along with the successful definitions in the "errors" response node.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:body (BulkUpdateSimNetworkPreference):

Returns

on success {:error, Tesla.Env.t} on failure

list_sim_card_data_usage_notifications(connection, opts \\ [])

@spec list_sim_card_data_usage_notifications(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.SimCardDataUsageNotificationCollection.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List SIM card data usage notifications Lists a paginated collection of SIM card data usage notifications. It enables exploring the collection using specific filters.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:filter_left_square_bracketsim_card_id_right_square_bracket (String.t): A valid SIM card ID.

Returns

on success {:error, Tesla.Env.t} on failure

post_validate_registration_codes(connection, create_sim_card_validate_request, opts \\ [])

@spec post_validate_registration_codes(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateSimCardValidateRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.SimCardRegistrationCodeValidations.t()}
  | {:error, Tesla.Env.t()}

Validate SIM cards registration codes It validates whether SIM card registration codes are valid or not.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_sim_card_validate_request (CreateSimCardValidateRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

remove_sim_card_public_ip(connection, id, opts \\ [])

@spec remove_sim_card_public_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SimCardActionResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Request removing a SIM card public IP This API removes an existing public IP from a SIM card. <br/><br/> The API will trigger an asynchronous operation called a SIM Card Action. The status of the SIM Card Action can be followed through the List SIM Card Action API.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

s_im_card_network_preferences_delete(connection, sim_card_id, opts \\ [])

@spec s_im_card_network_preferences_delete(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.DeleteSimCardNetworkPreferenceResponse.t()}
  | {:error, Tesla.Env.t()}

DELETE network preferences This API asynchronously removes the custom-defined network preferences settings. After this operation is done the Telnyx default settings, the same applied for an unaltered SIM card, will be in place.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_id (String.t): Identifies a SIM card.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

s_im_card_network_preferences_get(connection, sim_card_id, opts \\ [])

@spec s_im_card_network_preferences_get(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SimCardNetworkPreferenceWithOtaUpdatesResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Get network preferences It returns the network preferences currently applied in the SIM card.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_id (String.t): Identifies a SIM card.

opts (KeywordList): [optional] Optional parameters

:include_ota_updates (boolean()): It includes the associated OTA update objects in the response when present.

Returns

on success {:error, Tesla.Env.t} on failure

s_im_card_network_preferences_put(connection, sim_card_id, opts \\ [])

@spec s_im_card_network_preferences_put(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.PutSimCardNetworkPreferenceResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Set network preferences This API allows setting or updating a SIM card network preference. <br/><br/> Every SIM card has default network preferences defined on Telnyx. These preferences will determine how a SIMCard will connect to the network by considering a list of preferable operators.<br/><br/> There can be multiple scenarios where an operator can be preferred over another, for example, when a specific mobile operator can provide better network latency or better pricing.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_id (String.t): Identifies a SIM card.

opts (KeywordList): [optional] Optional parameters

:body (PutNetworkPreferenceRequest):

Returns

on success {:error, Tesla.Env.t} on failure

s_im_card_public_ip_get(connection, sim_card_id, opts \\ [])

@spec s_im_card_public_ip_get(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.SimCardPublicIpResponse.t()}
  | {:error, Tesla.Env.t()}

Get SIM card public IP definition It returns the public IP requested for a SIM card.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_id (String.t): Identifies a SIM card.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

set_sim_card_public_ip(connection, id, opts \\ [])

@spec set_sim_card_public_ip(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SimCardActionResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Request setting a SIM card public IP This API makes a SIM card reachable on the public internet by mapping a random public IP to the SIM card. <br/><br/> The API will trigger an asynchronous operation called a SIM Card Action. The status of the SIM Card Action can be followed through the List SIM Card Action API. <br/><br/> Setting a Public IP to a SIM Card incurs a charge and will only succeed if the account has sufficient funds.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_data_usage_notifications_delete(connection, id, opts \\ [])

@spec sim_card_data_usage_notifications_delete(
  Tesla.Env.client(),
  String.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.DeleteSimCardDataUsageNotificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Delete SIM card data usage notifications Delete the SIM Card Data Usage Notification.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_data_usage_notifications_get(connection, id, opts \\ [])

@spec sim_card_data_usage_notifications_get(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetSimCardDataUsageNotificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.Error.t()}
  | {:error, Tesla.Env.t()}

Get a single SIM card data usage notification Get a single SIM Card Data Usage Notification.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_data_usage_notifications_patch(connection, id, sim_card_data_usage_notification, opts \\ [])

@spec sim_card_data_usage_notifications_patch(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.SimCardDataUsageNotification.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.UpdateSimCardDataUsageNotificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Updates information for a SIM Card Data Usage Notification Updates information for a SIM Card Data Usage Notification.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

sim_card_data_usage_notification (SimCardDataUsageNotification):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_data_usage_notifications_post(connection, create_sim_card_data_usage_notification_request, opts \\ [])

@spec sim_card_data_usage_notifications_post(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateSimCardDataUsageNotificationRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateSimCardDataUsageNotificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a new SIM card data usage notification Creates a new SIM card data usage notification.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_sim_card_data_usage_notification_request (CreateSimCardDataUsageNotificationRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_delete(connection, id, opts \\ [])

@spec sim_card_delete(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.DeleteSimCardResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Deletes a SIM card The SIM card will be decommissioned, removed from your account and you will stop being charged.<br />The SIM card won't be able to connect to the network after the deletion is completed, thus making it impossible to consume data.<br/> Transitioning to the disabled state may take a period of time.</br> Until the transition is completed, the SIM card status will be disabling <code>disabling</code>.<br />In order to re-enable the SIM card, you will need to re-register it.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_disable(connection, id, opts \\ [])

@spec sim_card_disable(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SimCardActionResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Request a SIM card disable This API disables a SIM card, disconnecting it from the network and making it impossible to consume data.<br/> The API will trigger an asynchronous operation called a SIM Card Action. Transitioning to the disabled state may take a period of time. The status of the SIM Card Action can be followed through the List SIM Card Action API.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_enable(connection, id, opts \\ [])

@spec sim_card_enable(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SimCardActionResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Request a SIM card enable This API enables a SIM card, connecting it to the network and making it possible to consume data.<br/> To enable a SIM card, it must be associated with a SIM card group.<br/> The API will trigger an asynchronous operation called a SIM Card Action. Transitioning to the enabled state may take a period of time. The status of the SIM Card Action can be followed through the List SIM Card Action API.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_get(connection, id, opts \\ [])

@spec sim_card_get(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.GetSimCardResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Get SIM card Returns the details regarding a specific SIM card.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

:include_sim_card_group (boolean()): It includes the associated SIM card group object in the response when present.

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_register(connection, sim_card_registration, opts \\ [])

@spec sim_card_register(
  Tesla.Env.client(),
  TelnyxAPI.Model.SimCardRegistration.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.RegisterSimCardsResponse.t()} | {:error, Tesla.Env.t()}

Register SIM cards Register the SIM cards associated with the provided registration codes to the current user's account.<br/><br/> If <code>sim_card_group_id</code> is provided, the SIM cards will be associated with that group. Otherwise, the default group for the current user will be used.<br/><br/>

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_registration (SimCardRegistration):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_set_standby(connection, id, opts \\ [])

@spec sim_card_set_standby(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.SimCardActionResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Request setting a SIM card to standby The SIM card will be able to connect to the network once the process to set it to standby has been completed, thus making it possible to consume data.<br/> To set a SIM card to standby, it must be associated with SIM card group.<br/> The API will trigger an asynchronous operation called a SIM Card Action. Transitioning to the standby state may take a period of time. The status of the SIM Card Action can be followed through the List SIM Card Action API.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_card_update(connection, id, sim_card, opts \\ [])

@spec sim_card_update(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.SimCard.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.UpdateSimCardResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a SIM card Updates SIM card data

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Identifies the resource.

sim_card (SimCard):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

sim_cards_get(connection, opts \\ [])

@spec sim_cards_get(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.SearchSimCardsResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Get all SIM cards Get all SIM cards belonging to the user that match the given filters.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

:include_sim_card_group (boolean()): It includes the associated SIM card group object in the response when present.

:filter_left_square_bracketsim_card_group_id_right_square_bracket (String.t): A valid SIM card group ID.

:filter_left_square_brackettags_right_square_bracket ([String.t]): A list of SIM card tags to filter on.<br/><br/> If the SIM card contains <b><i>all</i></b> of the given <code>tags</code> they will be found.<br/><br/> For example, if the SIM cards have the following tags: <ul> <li><code>['customers', 'staff', 'test']</code> <li><code>['test']</code></li> <li><code>['customers']</code></li> </ul> Searching for <code>['customers', 'test']</code> returns only the first because it's the only one with both tags.<br/> Searching for <code>test</code> returns the first two SIMs, because both of them have such tag.<br/> Searching for <code>customers</code> returns the first and last SIMs.<br/>

:filter_left_square_bracketiccid_right_square_bracket (String.t): A search string to partially match for the SIM card's ICCID.

Returns

on success {:error, Tesla.Env.t} on failure

wireless_connectivity_logs_get(connection, sim_card_id, opts \\ [])

@spec wireless_connectivity_logs_get(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.WirelessConnectivityLogCollectionResponse.t()}
  | {:error, Tesla.Env.t()}

List wireless connectivity logs This API allows listing a paginated collection of Wireless Connectivity Logs associated with a SIM Card, for troubleshooting purposes.

Parameters

connection (TelnyxAPI.Connection): Connection to server

sim_card_id (String.t): Identifies a SIM card.

opts (KeywordList): [optional] Optional parameters

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure