TelnyxAPI.Api.Verify (telnyx_api v2.0.0)

API calls for all endpoints tagged Verify.

Functions

create_verification_call(connection, create_verification_request_call, opts \\ [])

@spec create_verification_call(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateVerificationRequestCall.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateVerificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Trigger a Call verification

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_verification_request_call (CreateVerificationRequestCall):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_verification_flashcall(connection, create_verification_request_flashcall, opts \\ [])

@spec create_verification_flashcall(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateVerificationRequestFlashcall.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateVerificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Trigger a Flash call verification

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_verification_request_flashcall (CreateVerificationRequestFlashcall):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_verification_psd2(connection, create_verification_request_psd2, opts \\ [])

@spec create_verification_psd2(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateVerificationRequestPsd2.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateVerificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Trigger a PSD2 verification

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_verification_request_psd2 (CreateVerificationRequestPsd2):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_verification_sms(connection, create_verification_request_sms, opts \\ [])

@spec create_verification_sms(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateVerificationRequestSms.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateVerificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Trigger a SMS verification

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_verification_request_sms (CreateVerificationRequestSms):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_verification_whatsapp(connection, create_verification_request_whatsapp, opts \\ [])

@spec create_verification_whatsapp(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateVerificationRequestWhatsapp.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.CreateVerificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Trigger a Whatsapp verification

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_verification_request_whatsapp (CreateVerificationRequestWhatsapp):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_verify_profile(connection, create_verify_profile_request, opts \\ [])

@spec create_verify_profile(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateVerifyProfileRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.VerifyProfileResponseDataWrapper.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Create a Verify profile Creates a new Verify profile to associate verifications with.

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_verify_profile_request (CreateVerifyProfileRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

delete_verify_profile(connection, verify_profile_id, opts \\ [])

@spec delete_verify_profile(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.VerifyProfileResponseDataWrapper.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Delete a Verify profile

Parameters

connection (TelnyxAPI.Connection): Connection to server

verify_profile_id (String.t): The identifier of the Verify profile to delete.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_verifications(connection, phone_number, opts \\ [])

@spec list_verifications(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.ListVerificationsResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List verifications by phone number

Parameters

connection (TelnyxAPI.Connection): Connection to server

phone_number (String.t): The phone number associated with the verifications to retrieve.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_verify_profiles(connection, opts \\ [])

@spec list_verify_profiles(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.ListVerifyProfilesResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

List all Verify profiles Gets a paginated list of Verify profiles.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketname_right_square_bracket (String.t):

:page_left_square_bracketsize_right_square_bracket (integer()):

:page_left_square_bracketnumber_right_square_bracket (integer()):

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_verification(connection, verification_id, opts \\ [])

@spec retrieve_verification(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.RetrieveVerificationResponse.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a verification

Parameters

connection (TelnyxAPI.Connection): Connection to server

verification_id (String.t): The identifier of the verification to retrieve.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_verify_profile(connection, verify_profile_id, opts \\ [])

@spec retrieve_verify_profile(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, TelnyxAPI.Model.VerifyProfileResponseDataWrapper.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Retrieve a Verify profile Gets a single Verify profile.

Parameters

connection (TelnyxAPI.Connection): Connection to server

verify_profile_id (String.t): The identifier of the Verify profile to retrieve.

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

update_verify_profile(connection, verify_profile_id, update_verify_profile_request, opts \\ [])

@spec update_verify_profile(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateVerifyProfileRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.VerifyProfileResponseDataWrapper.t()}
  | {:ok, TelnyxAPI.Model.Errors.t()}
  | {:error, Tesla.Env.t()}

Update a Verify profile

Parameters

connection (TelnyxAPI.Connection): Connection to server

verify_profile_id (String.t): The identifier of the Verify profile to update.

update_verify_profile_request (UpdateVerifyProfileRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

verify_verification_code(connection, phone_number, verify_verification_code_request, opts \\ [])

@spec verify_verification_code(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.VerifyVerificationCodeRequest.t(),
  keyword()
) ::
  {:ok, TelnyxAPI.Model.Errors.t()}
  | {:ok, TelnyxAPI.Model.VerifyVerificationCodeResponse.t()}
  | {:error, Tesla.Env.t()}

Submit a verification code. Deprecation notice: `verify_profile_id' is now a mandatory field. The sunset date for this change is 2022-04-02 00:00:00. ## Parameters - connection (TelnyxAPI.Connection): Connection to server - phone_number (String.t): The phone number associated with the verification code being verified. - verify_verification_code_request (VerifyVerificationCodeRequest): - opts (KeywordList): [optional] Optional parameters ## Returns {:ok, TelnyxAPI.Model.VerifyVerificationCodeResponse.t} on success {:error, Tesla.Env.t} on failure