TelnyxAPI.Api.ConferenceCommands (telnyx_api v2.0.0)

API calls for all endpoints tagged ConferenceCommands.

Functions

conference_dial_participant_in(connection, id, conference_call_request, opts \\ [])

@spec conference_dial_participant_in(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceCallRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Dial a new participant into a conference Dials a phone number and, when the call is answered, automatically joins them into the specified conference. Expected Webhooks: - call.hangup - call.answered - conference.participant.joined - conference.participant.left

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_call_request (ConferenceCallRequest): Dial Into Conference request object

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_hold_participants(connection, id, conference_hold_request, opts \\ [])

@spec conference_hold_participants(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceHoldRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Hold conference participants Hold a list of participants in a conference call

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_hold_request (ConferenceHoldRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_join(connection, id, join_conference_request, opts \\ [])

@spec conference_join(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.JoinConferenceRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Join a conference Join an existing call leg to a conference. Issue the Join Conference command with the conference ID in the path and the call_control_id of the leg you wish to join to the conference as an attribute. The conference can have up to a certain amount of active participants, as set by the max_participants parameter in conference creation request. Expected Webhooks: - conference.participant.joined - conference.participant.left

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

join_conference_request (JoinConferenceRequest): Join Conference request object

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_leave(connection, id, leave_conference_request, opts \\ [])

@spec conference_leave(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.LeaveConferenceRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Leave a conference Removes a call leg from a conference and moves it back to parked state. Expected Webhooks: - conference.participant.left

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

leave_conference_request (LeaveConferenceRequest): Leave Conference request object

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_mute_participants(connection, id, conference_mute_request, opts \\ [])

@spec conference_mute_participants(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceMuteRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Mute conference participants Mute a list of participants in a conference call

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_mute_request (ConferenceMuteRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_play_audio(connection, id, conference_play_request, opts \\ [])

@spec conference_play_audio(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferencePlayRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Play audio to conference participants Play audio to all or some participants on a conference call.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_play_request (ConferencePlayRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_speak_text(connection, id, conference_speak_request, opts \\ [])

@spec conference_speak_text(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceSpeakRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Speak text to conference participants Convert text to speech and play it to all or some participants.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Specifies the conference by id or name

conference_speak_request (ConferenceSpeakRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_start_recording(connection, id, start_recording_request, opts \\ [])

@spec conference_start_recording(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.StartRecordingRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Conference recording start Start recording the conference. Recording will stop on conference end, or via the Stop Recording command. Expected Webhooks: - conference.recording.saved

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Specifies the conference to record by id or name

start_recording_request (StartRecordingRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_stop_audio(connection, id, conference_stop_request, opts \\ [])

@spec conference_stop_audio(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceStopRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Stop audio being played on the conference Stop audio being played to all or some participants on a conference call.

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_stop_request (ConferenceStopRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_stop_recording(connection, id, stop_recording_request, opts \\ [])

@spec conference_stop_recording(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.StopRecordingRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Conference recording stop Stop recording the conference. Expected Webhooks: - conference.recording.saved

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Specifies the conference to stop the recording for by id or name

stop_recording_request (StopRecordingRequest): Stop recording conference request

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_unhold_participants(connection, id, conference_unhold_request, opts \\ [])

@spec conference_unhold_participants(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceUnholdRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Unhold conference participants Unhold a list of participants in a conference call

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_unhold_request (ConferenceUnholdRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_unmute_participants(connection, id, conference_unmute_request, opts \\ [])

@spec conference_unmute_participants(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.ConferenceUnmuteRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Unmute conference participants Unmute a list of participants in a conference call

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

conference_unmute_request (ConferenceUnmuteRequest):

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

conference_update(connection, id, update_conference_request, opts \\ [])

@spec conference_update(
  Tesla.Env.client(),
  String.t(),
  TelnyxAPI.Model.UpdateConferenceRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceCommandResponse.t()}
  | {:error, Tesla.Env.t()}

Update conference participant Update conference participant supervisor_role

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id or name

update_conference_request (UpdateConferenceRequest): Update Conference request object

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

create_conference(connection, create_conference_request, opts \\ [])

@spec create_conference(
  Tesla.Env.client(),
  TelnyxAPI.Model.CreateConferenceRequest.t(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceResponse.t()}
  | {:error, Tesla.Env.t()}

Create conference Create a conference from an existing call leg using a call_control_id and a conference name. Upon creating the conference, the call will be automatically bridged to the conference. Conferences will expire after all participants have left the conference or after 4 hours regardless of the number of active participants. Expected Webhooks: - conference.created - conference.participant.joined - conference.participant.left - conference.ended - conference.recording.saved - conference.floor.changed

Parameters

connection (TelnyxAPI.Connection): Connection to server

create_conference_request (CreateConferenceRequest): Create a conference

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure

list_conference_participants(connection, conference_id, opts \\ [])

@spec list_conference_participants(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListParticipantsResponse.t()}
  | {:error, Tesla.Env.t()}

List conference participants Lists conference participants

Parameters

connection (TelnyxAPI.Connection): Connection to server

conference_id (String.t): Uniquely identifies the conference by id

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketmuted_right_square_bracket (boolean()): If present, participants will be filtered to those who are/are not muted

:filter_left_square_bracketon_hold_right_square_bracket (boolean()): If present, participants will be filtered to those who are/are not put on hold

:filter_left_square_bracketwhispering_right_square_bracket (boolean()): If present, participants will be filtered to those who are whispering or are not

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

list_conferences(connection, opts \\ [])

@spec list_conferences(
  Tesla.Env.client(),
  keyword()
) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ListConferencesResponse.t()}
  | {:error, Tesla.Env.t()}

List conferences Lists conferences. Conferences are created on demand, and will expire after all participants have left the conference or after 4 hours regardless of the number of active participants. Conferences are listed in descending order by expires_at.

Parameters

connection (TelnyxAPI.Connection): Connection to server

opts (KeywordList): [optional] Optional parameters

:filter_left_square_bracketname_right_square_bracket (String.t): If present, conferences will be filtered to those with a matching name attribute. Matching is case-sensitive

:filter_left_square_bracketstatus_right_square_bracket (String.t): If present, conferences will be filtered by status.

:page_left_square_bracketnumber_right_square_bracket (integer()): The page number to load

:page_left_square_bracketsize_right_square_bracket (integer()): The size of the page

Returns

on success {:error, Tesla.Env.t} on failure

retrieve_conference(connection, id, opts \\ [])

@spec retrieve_conference(Tesla.Env.client(), String.t(), keyword()) ::
  {:ok, nil}
  | {:ok, TelnyxAPI.Model.ConferenceResponse.t()}
  | {:error, Tesla.Env.t()}

Retrieve a conference Retrieve an existing conference

Parameters

connection (TelnyxAPI.Connection): Connection to server

id (String.t): Uniquely identifies the conference by id

opts (KeywordList): [optional] Optional parameters

Returns

on success {:error, Tesla.Env.t} on failure